### 2021-07-23 #3530 @whedon
[PRE REVIEW]: Manopt.jl: Optimization on Manifolds in Julia
**Submitting author:** @kellertuer (<a href="http://orcid.org/0000-0001-8342-7218">Ronny Bergmann</a>)
**Repository:** <a href="https://github.com/JuliaManifolds/Manopt.jl" target ="_blank">https://github.com/JuliaManifolds/Manopt.jl</a>
**Version:** 0.3.11
