### 2021-08-03 a65ce39
use default retraction/vector transport in solvers
