+++ src/solvers/ChambollePock.jl
@@ -47,9 +47,9 @@ For more details on the algorithm, see[^BergmannHerzogSilvaLouzeiroTenbrinckVida
 * `stopping_criterion` – (`stopAtIteration(100)`) a [`StoppingCriterion`](@ref)
 * `update_primal_base` – (`missing`) function to update `m` (identity by default/missing)
 * `update_dual_base` – (`missing`) function to update `n` (identity by default/missing)
-* `retraction_method` – (`ExponentialRetraction()`) the rectraction to use
-* `inverse_retraction_method` - (`LogarithmicInverseRetraction()`) an inverse retraction to use.
-* `vector_transport_method` - (`ParallelTransport()`) a vector transport to use
+* `retraction_method` – (`default_retraction_method(M)`) the rectraction to use
+* `inverse_retraction_method` - (`default_inverse_retraction_method(N)`) an inverse retraction to use.
+* `vector_transport_method` - (`default_vector_transport_method(M)`) a vector transport to use
 
 [^BergmannHerzogSilvaLouzeiroTenbrinckVidalNunez2020]:
     > R. Bergmann, R. Herzog, M. Silva Louzeiro, D. Tenbrinck, J. Vidal-Núñez:
@@ -124,9 +124,9 @@ function ChambollePock!(
     stopping_criterion::StoppingCriterion=StopAfterIteration(200),
     update_primal_base::Union{Function,Missing}=missing,
     update_dual_base::Union{Function,Missing}=missing,
-    retraction_method::RM=ExponentialRetraction(),
-    inverse_retraction_method::IRM=LogarithmicInverseRetraction(),
-    vector_transport_method::VTM=ParallelTransport(),
+    retraction_method::RM=default_retraction_method(M),
+    inverse_retraction_method::IRM=default_inverse_retraction_method(N),
+    vector_transport_method::VTM=default_vector_transport_method(M),
     variant=ismissing(Λ) ? :exact : :linearized,
     return_options=false,
     kwargs...,
