+++ Project.toml
@@ -25,7 +25,7 @@ ColorTypes = "0.9.1, 0.10, 0.11"
 Colors = "0.11.2, 0.12"
 DataStructures = "0.17, 0.18"
 Manifolds = "0.5"
-ManifoldsBase = "0.11, 0.12"
+ManifoldsBase = "0.12.3"
 Requires = "0.5, 1"
 StaticArrays = "0.12, 1.0"
 julia = "1.4"
