### 2021-08-04 e0ec985
Use default retraction and vector transport method in solvers (#91)
* use default retraction/vector transport in solvers
* fix typo
* increment version number
* default inverse retraction for Chambolle-Pock
* import default operations from ManifoldsBase

Co-authored-by: Nick Dewaele <nick.dewaele@kuleuven.be>
