+++ src/solvers/conjugate_gradient_descent.jl
@@ -35,7 +35,7 @@ They all compute ``β_k`` such that this algorithm updates the search direction 
   [`ConjugateGradientDescentOptions`](@ref) `o` and `i` is the current iterate.
 * `evaluation` – ([`AllocatingEvaluation`](@ref)) specify whether the gradient works by allocation (default) form `gradF(M, x)`
   or [`MutatingEvaluation`](@ref) in place, i.e. is of the form `gradF!(M, X, x)`.
-* `retraction_method` - (`ExponentialRetraction`) a retraction method to use, by default the exponential map
+* `retraction_method` - (`default_retraction_method(M`) a retraction method to use.
 * `return_options` – (`false`) – if actiavated, the extended result, i.e. the
     complete [`Options`](@ref) re returned. This can be used to access recorded values.
     If set to false (default) just the optimal value `x_opt` if returned
@@ -43,7 +43,7 @@ They all compute ``β_k`` such that this algorithm updates the search direction 
   search direction. The default is a constant step size 1.
 * `stopping_criterion` : (`stopWhenAny( stopAtIteration(200), stopGradientNormLess(10.0^-8))`)
   a function indicating when to stop.
-* `vector_transport_method` – (`ParallelTransport()`) vector transport method to transport
+* `vector_transport_method` – (`default_vector_transport_method(M)`) vector transport method to transport
   the old descent direction when computing the new descent direction.
 
 # Output
@@ -83,11 +83,11 @@ function conjugate_gradient_descent!(
     x;
     coefficient::DirectionUpdateRule=SteepestDirectionUpdateRule(),
     stepsize::Stepsize=ConstantStepsize(1.0),
-    retraction_method::AbstractRetractionMethod=ExponentialRetraction(),
+    retraction_method::AbstractRetractionMethod=default_retraction_method(M),
     stopping_criterion::StoppingCriterion=StopWhenAny(
         StopAfterIteration(500), StopWhenGradientNormLess(10^(-8))
     ),
-    transport_method=ParallelTransport(),
+    transport_method=default_vector_transport_method(M),
     return_options=false,
     kwargs...,
 ) where {TF,TDF}
