### 2021-08-05 #3530 @dpsanders
👋 @kellertuer: Thanks for submitting to JOSS! This looks like a nice package and paper.

Here are some initial comments:

- Perhaps `ManifoldOptimization.jl` would be a better name for the package?

- Could you please check over the paper for grammar. In particular the paragraph ending on line 48 ends in the middle of a sentence.

- Maybe include output and a figure in the paper for the example, if that is possible?
