### 2021-08-05 #3530 @kellertuer
Thanks for the first feedback!

> * Perhaps `ManifoldOptimization.jl` would be a better name for the package?

Thanks for the idea. I know that in Julia abbreviations in names are not that common. The reason for the shorter name stems from the fact that manopt (small letter, no ending) is a Matlab package with the same goal since 2013 and pymanopt similarly a variant in Python. So to keep this packages the name is `Manopt.jl`.

Another point is that `ManifoldOptimization` might be misleading, that it might bear the impression one optimises _a_ manifold (parametrised on some way, for example a surface) and not _optimizse on_ a (itself fixed) manifold. Then, `OptimzationOnManifolds.jl` would be more correct, but also not a nice package name.

> * Could you please check over the paper for grammar. In particular the paragraph ending on line 48 ends in the middle of a sentence.

Oh, sorry for that, I will carefully read the paper again tomorrow (CEST here).

> * Maybe include output and a figure in the paper for the example, if that is possible?

The output for the result of the first part would just be a vector of length 3 – maybe a little poring?
For the second part that can be done, sure. I think that were 17 iterations, then I would maybe even introduce a “subsampling” of only every third or fourth iterate to not produce 19 lines of output.

The input data and the mean can surely also be shown in a figure. I will add that also tomorrow, no problem.
Adding even the iterates (from the `get_record` to the figure might be again a little boring, since it converges quite fast and most points will look identical to the result on the sphere.
