### 2021-08-06 14062d0
Revise paper and extend the example by some output and a figure.
