+++ joss/paper.md
@@ -20,46 +20,46 @@ bibliography: bibliography.bib
 # Summary
 
 [`Manopt.jl`](https://manoptjl.org) provides a set of optimization algorithms for optimization problems given on a Riemannian manifold $\mathcal M$.
-Based on a generic optimization framework together with the interface `ManifoldsBase.jl` for Riemannian manifolds, classical and recently developed methods are provided in an efficient implementation. Algorithms include the least requiring Particle Swarm and Nelder Mead algorithms as well as a classical gradient or stochastic gradient descent. Furthermore, quasi Newton methods like a Riemannian L-BFGS [@HuangGallivanAbsil:2015:1] and nonsmooth optimization algorithms like a Cyclic Proximal Point Algorithm [@Bacak:2014:1], Douglas-Rachford [@BergmannPerschSteidl:2016:1] and Chambolle-Pock algorithm [@BergmannHerzogSilvaLouzeiroTenbrinckVidalNunez:2021:1] are provided together with several basic cost functions, gradients and proximal maps as well as debug and record capabilities.
+Based on a generic optimization framework together with the interface `ManifoldsBase.jl` for Riemannian manifolds, classical and recently developed methods are provided in an efficient implementation. Algorithms include the derivative free Particle Swarm and Nelder Mead algorithms as well as a classical gradient, conjugate gradient and stochastic gradient descent. Furthermore, quasi Newton methods like a Riemannian L-BFGS [@HuangGallivanAbsil:2015:1] and nonsmooth optimization algorithms like a Cyclic Proximal Point Algorithm [@Bacak:2014:1], a (parallel) Douglas-Rachford algorithm [@BergmannPerschSteidl:2016:1] and a Chambolle-Pock algorithm [@BergmannHerzogSilvaLouzeiroTenbrinckVidalNunez:2021:1] are provided together with several basic cost functions, gradients and proximal maps as well as debug and record capabilities.
 
 # Statement of Need
 
 In many applications and optimization tasks, nonlinear data appears naturally.
-For example when data on the sphere is measured, diffusion data can be captures as a signal or multivariate data of symmetric positive definite matrices or orientations like they appear for electron backscattered diffraction (EBSD) data. Another example are fixed rank matrices, appearing in dictionary learning.
-Working on these data, for example doing data interpolation, data approximation, denoising, inpainting, or performing matrix completion, can be phrased as an optimization problem
+For example when data on the sphere is measured, diffusion data can be captured as a signal or even multivariate data of symmetric positive definite matrices, and orientations like they appear for electron backscattered diffraction (EBSD) data. Another example are fixed rank matrices, appearing in dictionary learning.
+Working on these data, for example doing data interpolation, data approximation, denoising, inpainting, or performing matrix completion, can usually phrased as an optimization problem
 
 $$ \text{Minimize}\quad f(x) \quad \text{where } x\in\mathcal M, $$
 
 where the optimization problem is phrased on a Riemannian manifold $\mathcal M$.
 
-A main challenge of these algorithms is that, compared to the (classical) Euclidean case, there is no addition available. For example on the unit sphere $\mathbb S^2$ of unit vectors in $\mathbb R^3$, adding two vectors of unit lengths yields a point that is not of unit norm.
-the resolution is to generalize the notion of a shortest path from the straight line to what is called a (shortest) geodesic, or acceleration free curves.
+A main challenge of these algorithms is that, compared to the (classical) Euclidean case, there is no addition available. For example on the unit sphere $\mathbb S^2$ of unit vectors in $\mathbb R^3$, adding two vectors of unit lengths yields a vector that is not of unit norm.
[…]
