+++ joss/bibliography.bib
@@ -144,7 +144,7 @@
     JOURNALTITLE = {Association for Computing Machinery. Transactions on Mathematical Software},
     NUMBER       = {4},
     PAGES        = {Art. 43,21},
-    TITLE        = {ROPTLIB: an object-oriented C++ library for optimization on Riemannian manifolds},
+    TITLE        = {ROPTLIB: an object-oriented {C++} library for optimization on Riemannian manifolds},
     VOLUME       = {44}
 }
 
@@ -170,4 +170,15 @@
     AUTHOR    = {P.-A. Absil and C.G. Baker and K.A. Gallivan},
     TITLE     = {Trust-Region Methods on Riemannian Manifolds},
     JOURNAL   = {Foundations of Computational Mathematics}
+}
+
+@article{Karcher:1977:1,
+    AUTHOR       = {Karcher, H.},
+    DATE         = {1977},
+    DOI          = {10.1002/cpa.3160300502},
+    JOURNALTITLE = {Communications on Pure and Applied Mathematics},
+    NUMBER       = {5},
+    PAGES        = {509--541},
+    TITLE        = {Riemannian center of mass and mollifier smoothing},
+    VOLUME       = {30}
 }
