### 2021-08-06 #3530 @kellertuer
I addressed the points from the second and third point in the most recent commit.
