### 2021-08-06 078a500
change Minimize to an arg min.
