+++ src/solvers/primal_dual_semismooth_Newton.jl
@@ -1,0 +1,315 @@
+@doc raw"""
+ChambollePock(M, N, cost, x0, ξ0, m, n, prox_F, prox_G_dual, forward_operator, adjoint_DΛ)
+
+Perform the Riemannian Chambolle–Pock algorithm.
+
+Given a `cost` function $\mathcal E\colon\mathcal M \to ℝ$ of the form
+```math
+\mathcal E(x) = F(x) + G( Λ(x) ),
+```
+where $F\colon\mathcal M \to ℝ$, $G\colon\mathcal N \to ℝ$,
+and $\Lambda\colon\mathcal M \to \mathcal N$. The remaining input parameters are
+
+* `x,ξ` primal and dual start points $x\in\mathcal M$ and $\xi\in T_n\mathcal N$
+* `m,n` base points on $\mathcal M$ and $\mathcal N$, respectively.
+* `forward_operator` the operator $Λ(⋅)$ or its linearization $DΛ(⋅)[⋅]$, depending on whether `:exact` or `:linearized` is chosen.
+* `adjoint_linearized_operator` the adjoint $DΛ^*$ of the linearized operator $DΛ(m)\colon T_{m}\mathcal M \to T_{Λ(m)}\mathcal N$
+* `prox_F, prox_G_Dual` the proximal maps of $F$ and $G^\ast_n$
+
+By default, this performs the exact Riemannian Chambolle Pock algorithm, see the opional parameter
+`DΛ` for ther linearized variant.
+
+For more details on the algorithm, see[^BergmannHerzogSilvaLouzeiroTenbrinckVidalNunez2020].
+
+# Optional Parameters
+
+* `acceleration` – (`0.05`)
+* `dual_stepsize` – (`1/sqrt(8)`) proximnal parameter of the primal prox
+* `Λ` (`missing`) the exact operator, that is required if the forward operator is linearized;
+`missing` indicates, that the forward operator is exact.
+* `primal_stepsize` – (`1/sqrt(8)`) proximnal parameter of the dual prox
+* `relaxation` – (`1.`)
+* `relax` – (`:primal`) whether to relax the primal or dual
+* `variant` - (`:exact` if `Λ` is missing, otherwise `:linearized`) variant to use.
+Note that this changes the arguments the `forward_operator` will be called.
+* `stopping_criterion` – (`stopAtIteration(100)`) a [`StoppingCriterion`](@ref)
+* `update_primal_base` – (`missing`) function to update `m` (identity by default/missing)
+* `update_dual_base` – (`missing`) function to update `n` (identity by default/missing)
+* `retraction_method` – (`ExponentialRetraction()`) the rectraction to use
+* `inverse_retraction_method` - (`LogarithmicInverseRetraction()`) an inverse retraction to use.
+* `vector_transport_method` - (`ParallelTransport()`) a vector transport to use
+
+[^BergmannHerzogSilvaLouzeiroTenbrinckVidalNunez2020]:
+> R. Bergmann, R. Herzog, M. Silva Louzeiro, D. Tenbrinck, J. Vidal-Núñez:
+> _Fenchel Duality Theory and a Primal-Dual Algorithm on Riemannian Manifolds_,
+> arXiv: [1908.02022](http://arxiv.org/abs/1908.02022)
+> accepted for publication in Foundations of Computational Mathematics
+"""
[…]
