+++ src/plans/higher_order_primal_dual_plan.jl
@@ -1,0 +1,576 @@
+# TODO: change documentation
+# DONE: addapted struct and functions
+@doc raw"""
+    PrimalDualProblem {mT <: Manifold, nT <: Manifold} <: PrimalDualProblem} <: Problem
+
+Describes a Problem for the linearized Chambolle-Pock algorithm.
+
+# Fields
+
+* `M`, `N` – two manifolds $\mathcal M$, $\mathcal N$
+* `cost` $F + G(Λ(⋅))$ to evaluate interims cost function values
+* `forward_oprator` the operator for the forward operation in the algorthm, either $Λ$ (exact) or $DΛ$ (linearized).
+* `linearized_adjoint_operator` The adjoint differential $(DΛ)^* \colon \mathcal N \to T\mathcal M$
+* `prox_F` the proximal map belonging to $f$
+* `prox_G_dual` the proximal map belonging to $g_n^*$
+* `Λ` – (`fordward_operator`) for the linearized variant, this has to be set to the exact forward operator.
+  This operator is required in several variants of the linearized algorithm.
+  Since the exact variant is the default, `Λ` is by default set to `forward_operator`.
+
+# Constructor
+
+    LinearizedPrimalDualProblem(M, N, cost, prox_F, prox_G_dual, forward_operator, adjoint_linearized_operator,Λ=forward_operator)
+
+"""
+mutable struct PrimalDualSemismoothNewtonProblem{mT<:Manifold,nT<:Manifold} <: Problem
+    M::mT
+    N::nT
+    cost::Function
+    prox_F::Function
+    diff_prox_F::Function
+    prox_G_dual::Function
+    diff_prox_G_dual::Function
+    forward_operator::Function
+    adjoint_linearized_operator::Function
+    Λ::Union{Function,Missing}
+end
+function PrimalDualSemismoothNewtonProblem(
+    M::mT,
+    N::nT,
+    cost,
+    prox_F,
+    diff_prox_F,
+    prox_G_dual,
+    diff_prox_G_dual,
+    forward_operator,
+    adjoint_linearized_operator,
+    Λ=forward_operator,
+) where {mT<:Manifold,nT<:Manifold}
+    return PrimalDualSemismoothNewtonProblem{mT,nT}(
+        M, N, cost, prox_F, prox_G_dual, forward_operator, adjoint_linearized_operator, Λ
+    )
+end
+
+# TODO: adapt documentation
+# TODO: do we actually need this here as well? Also in ChambollePock
+@doc raw"""
+    PrimalDualOptions
+
+A general type for all primal dual based options to be used within primal dual
+based algorithms
+"""
+abstract type PrimalDualOptions <: Options end
+
+@doc raw"""
+    ChambollePockOptions <: PrimalDualOptions
+
+stores all options and variables within a linearized or exact Chambolle Pock.
+The following list provides the order for the constructor, where the previous iterates are
+initialized automatically and values with a default may be left out.
+
+* `m` - base point on $ \mathcal M $
+* `n` - base point on $ \mathcal N $
+* `x` - an initial point on $x^{(0)} \in \mathcal M$ (and its previous iterate)
+* `ξ` - an initial tangent vector $\xi^{(0)}\in T^*\mathcal N$ (and its previous iterate)
+* `xbar` - the relaxed iterate used in the next dual update step (when using `:primal` relaxation)
+* `ξbar` - the relaxed iterate used in the next primal update step (when using `:dual` relaxation)
+* `Θ` – factor to damp the helping $\tilde x$
[…]
