+++ src/plans/plan.jl
@@ -11,6 +11,8 @@ include("gradient_plan.jl")
 include("hessian_plan.jl")
 include("proximal_plan.jl")
 include("subgradient_plan.jl")
+include("higher_order_primal_dual_plan.jl")
+
 include("particle_swarm_plan.jl")
 include("primal_dual_plan.jl")
 include("stochastic_gradient_plan.jl")
