+++ pluto/AutomaticDifferentiation.jl
@@ -1,0 +1,620 @@
+### A Pluto.jl notebook ###
+# v0.16.0
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ af02696b-a17e-4ec6-a81e-82d0469f4b71
+using ReverseDiff, Manifolds, Manopt, Test, Random, LinearAlgebra
+
+# ╔═╡ 0213d26a-18ac-11ec-03fd-ada5992bcea8
+md"""
+# Using (Euclidean) AD in Manopt.jl
+"""
+
+# ╔═╡ f3bc91ee-5871-4cba-ac89-190deb71ad0f
+md"""
+Since Manifolds.jl 0.6.7 the support of automatic differentiation support has been extended. With [ForwardDiff.jl](https://juliadiff.org/ForwardDiff.jl/stable/) or [ReverseDiff.jl](https://juliadiff.org/ReverseDiff.jl/) you can now use automatic differentiation on manifolds for embedded manifolds.
+"""
+
+# ╔═╡ 77769eab-54dd-41dc-8125-0382e5ef0bf1
+md"""
+## A Short Theory recap
+
+Let ``\tilde f\colon\mathbb R^m \to \mathbb R`` be a function un the embedding of an ``n``-dimensional manifold ``\mathcal M \subset \mathbb R^m`` and ``f\colon \mathcal M \to \mathbb R`` denote the restriction of ``\tilde f`` to the manifold ``\mathcal M``.
+
+Since we can use the push forward of the embedding to also embed the tangent space ``T_p\mathcal M``, ``p\in \mathcal M``, we can similarly obtain the differential ``Df(p)\colon T_p\mathcal M \to \mathbb R`` by restricting the differential ``D\tilde f(p)`` to the tangent space.
+
+If both ``T_p\mathcal M`` and ``T_p\mathcal R^m`` have the same inner product, or in other words the manifold is isometrically embedded in ``R^m`` (like for example the sphere ``\mathbb S^n\subset\mathbb R^{m+1}`` then this restriction of the differential directly translates to a projection of the gradient, i.e.
+
+```math
+\operatorname{grad}f(p) = \operatorname{Proj}_{T_p\mathcal M}(\operatorname{grad} \tilde f(p))
+```
+
+More generally we might have to take a change of the metric into account, i.e.
+
+```math
+\langle  \operatorname{Proj}_{T_p\mathcal M}(\operatorname{grad} \tilde f(p)), X \rangle
+= Df(p)[X] = g_p(\operatorname{grad}f(p), X)
+```
+
[…]
