### 2021-09-20 7b8853f
Started a section on finite differences.
