### 2021-09-20 d470ac0
Finishes the text on finite differences.
