### 2021-09-20 #3530 @dpsanders
👋 @matbesancon and @krystophny: Would you be available and willing to review this submission for [JOSS](https://joss.theoj.org/), the Journal of Open Source Software? Reviews are focused on the software and take place via GitHub issues. Thanks!
