### 2021-09-22 #3530 @kellertuer
@dpsanders we could ask Nicolas Boumal (nicolasboumal), developer of Manopt (Matlab, same philosophy – we even share the logo – but mostly different architecture of his package) or Sebastian Weichwald (sweichwald), one of the developers of pymanopt, which is a python adaption of manopt that focuses on AD backends.
Both are – as far as I saw – not in the JOSS list from above.
