### 2021-09-22 b573493
Continued to write about the theory.
