+++ pluto/AutomaticDifferentiation.jl
@@ -1,5 +1,5 @@
 ### A Pluto.jl notebook ###
-# v0.16.0
+# v0.16.1
 
 using Markdown
 using InteractiveUtils
@@ -22,9 +22,23 @@ md"""
 In general there is currenlty two ways to get the gradient of a function ``f\colon\mathcal M \to ℝ`` on a Riemannian manifold ``\mathcal M``.
 """
 
+# ╔═╡ d9be6c2f-65fd-4685-9005-da22bf985e28
+md"""
+In this Notebook we will take a look at a few possibilities to approximate or derive the gradient of a function ``f:\mathcal M \to ℝ`` on a Riemannian manifold, without computing it yourself. There is mainly two different philosophies:
+
+1. Working _instrinsically_, i.e. stay on the manifold and in the tangent spaces. Here, we will consider approximating the gradient by forward differences.
+
+2. Working in an embedding – there we can use all tools from functions on Euclidean spaces – finite differences or automatic differenciation – and then compute the corresponding Riemannian gradient from there.
+
+Let's first load all packages we need.
+"""
+
+# ╔═╡ f88b15de-cec6-4bc8-9b68-2a407b5aeded
+
+
 # ╔═╡ 18d7459f-eed6-489b-a096-ac77ccd781af
 md"""
-## Forward Differences
+## 1. (Intrinsic) Forward Differences
 
 A first idea is to generalise (multivariate) finite differences to Riemannian manifolds. Let ``X_1,\ldots,X_d ∈ T_p\mathcal M`` denote an orthonormal basis of the tangent space ``T_p\mathcal M`` at the point ``p∈\mathcal M`` on the Riemannian manifold.
 
@@ -66,11 +80,53 @@ This is the first variant we can use. An advantage is, that it is _intrinsic_ in
 """
 
 # ╔═╡ 9a030ac6-1f44-4fa6-8bc9-1c0278e97fe2
+md""" ### An Example: The Rayleigh Quotient
+
+The Rayleigh quotient is concerned with finding Eigenvalues (and Eigenvectors) of a symmetric matrix $A\in ℝ^{(n+1)×(n+1)}$. The optimisation problem reads
+
+```math
+F\colon ℝ^{n+1} \to ℝ,\quad F(\mathbf x) = \frac{\mathbf x^\mathrm{T}A\mathbf x}{\mathbf x^\mathrm{T}\mathbf x}
+```
+
+Minimizing this function yields the smallest eigenvalue ``\lambda_1`` as a value and the corresponding minimizer ``\mathbf x^*`` is a corresponding eigenvector.
+
+Since the length of an eigenvector is irrelevant, there is an ambiguity in the cost function. It can be better phrased on the sphere ``𝕊^n`` of unit vectors in ``\mathbb R^{n+1}``, i.e.
+
+```math
[…]
