### 2021-09-22 70cd81f
First approach to use RiemannianONBDiff
