+++ pluto/AutomaticDifferentiation.jl
@@ -1,11 +1,17 @@
 ### A Pluto.jl notebook ###
-# v0.16.1
+# v0.16.0
 
 using Markdown
 using InteractiveUtils
 
-# ╔═╡ 41c204dd-6e4e-4a70-8f06-209a469e0680
-using ReverseDiff, Manifolds, Manopt, Test, Random, LinearAlgebra
+# ╔═╡ b0769dfa-28cf-440e-9ba2-1ef488f171a9
+begin # dev mode – use global env on my machine
+    import Pkg
+    # careful: this is _not_ a reproducible environment
+    # activate the global environment
+    Pkg.activate()
+   using ReverseDiff, Manifolds, Manopt, Test, Random, LinearAlgebra, FiniteDiff
+end
 
 # ╔═╡ 0213d26a-18ac-11ec-03fd-ada5992bcea8
 md"""
@@ -34,7 +40,7 @@ Let's first load all packages we need.
 """
 
 # ╔═╡ f88b15de-cec6-4bc8-9b68-2a407b5aeded
-
+# using ReverseDiff, Manifolds, Manopt, Test, Random, LinearAlgebra
 
 # ╔═╡ 18d7459f-eed6-489b-a096-ac77ccd781af
 md"""
@@ -115,14 +121,41 @@ begin
 	nothing
 end
 
-# ╔═╡ 2e33de5e-ffaa-422a-91d9-61f588ed1211
+# ╔═╡ 41c204dd-6e4e-4a70-8f06-209a469e0680
+f1(p) = p'*A'p
 
+# ╔═╡ 2e33de5e-ffaa-422a-91d9-61f588ed1211
+gradf1(p) = 2*(A*p + p*p'*A*p) 
 
 # ╔═╡ bbd9a010-1981-45b3-bf7d-c04bcd2c2128
+md"""Manifolds provides a finite difference scheme"""
 
+# ╔═╡ 0c823b57-a009-44b7-9901-4b1a1ed7e103
+finite_diff = Manifolds.FiniteDiffBackend(Val(:forward))
+
+# ╔═╡ 47b536ea-cd9c-4083-bbd0-f69904a1307d
+
+
+# ╔═╡ 08456b40-74ec-4319-93e7-130b5cf70ac3
+r_backend = RiemannianONBDiffBackend(finite_diff, ExponentialRetraction(), LogarithmicInverseRetraction(), DefaultOrthonormalBasis())
 
 # ╔═╡ 12327b62-7e79-4381-b6a7-f85b08a8251b
+gradf1_FD(p) = Manifolds.gradient(M, f1, p, r_backend)
 
+# ╔═╡ 07f9a630-e53d-45ea-b109-3d4de190723d
+begin
+	p = zeros(n+1)
+	p[1] = 1.0
+	X1 = gradf1(p)
+	X2 = gradf1_FD(p)
+	[X1-X2]
+end
+
+# ╔═╡ 8e5f677d-dafa-49b9-b678-3f129be31dcf
+X1
+
+# ╔═╡ ab5faa5f-1394-40c2-8c81-0e0f5449cd72
+X2
 
 # ╔═╡ 77769eab-54dd-41dc-8125-0382e5ef0bf1
 md"""
@@ -164,7 +197,7 @@ f̃(x) = x'*A*x/(x'*x);
 md"The cost function is the same by restriction"
 
 # ╔═╡ c1341fef-adec-4574-a642-a1a8a9c1fee5
-f(p) = f̃(p);
+f2(p) = f̃(p);
 
 # ╔═╡ 0818a62f-1bef-44f7-a33f-1ab0054e853c
 md"The gradient is now computed combining our gradient scheme with ReverseDiff."
[…]
