+++ pluto/AutomaticDifferentiation.jl
@@ -114,7 +114,7 @@ so we can compare it to the approximation by finite differences.
 # ╔═╡ 19747159-d383-4547-9315-0ed2494904a6
 begin
 	Random.seed!(42)
-	n = 20
+	n = 200
 	A = randn(n+1,n+1)
 	A = Symmetric(A)
 	M = Sphere(n)
@@ -128,16 +128,10 @@ f1(p) = p'*A'p
 gradf1(p) = 2*(A*p - p*p'*A*p) 
 
 # ╔═╡ bbd9a010-1981-45b3-bf7d-c04bcd2c2128
-md"""Manifolds provides a finite difference scheme"""
-
-# ╔═╡ 0c823b57-a009-44b7-9901-4b1a1ed7e103
-finite_diff = Manifolds.FiniteDiffBackend(Val(:forward))
-
-# ╔═╡ 47b536ea-cd9c-4083-bbd0-f69904a1307d
-
+md"""Manifolds provides a finite difference scheme in Tangent spaces, that you can introduce to use an existing framework (if the wrapper is implemented) form Euclidean space. Here we use `FiniteDiff.jl`."""
 
 # ╔═╡ 08456b40-74ec-4319-93e7-130b5cf70ac3
-r_backend = RiemannianONBDiffBackend(finite_diff, ExponentialRetraction(), LogarithmicInverseRetraction(), DefaultOrthonormalBasis())
+r_backend = Manifolds.TangentDiffBackend(Manifolds.FiniteDiffBackend())
 
 # ╔═╡ 12327b62-7e79-4381-b6a7-f85b08a8251b
 gradf1_FD(p) = Manifolds.gradient(M, f1, p, r_backend)
@@ -148,14 +142,11 @@ begin
 	p[1] = 1.0
 	X1 = gradf1(p)
 	X2 = gradf1_FD(p)
-	[X1-X2]
+	norm(M, p, X1-X2)
 end
 
 # ╔═╡ 8e5f677d-dafa-49b9-b678-3f129be31dcf
-is_vector(M, p, X1)
-
-# ╔═╡ ab5faa5f-1394-40c2-8c81-0e0f5449cd72
-is_vector(M, p, X2)
+md"We obtain quite a good approximation of the gradient."
 
 # ╔═╡ 77769eab-54dd-41dc-8125-0382e5ef0bf1
 md"""
@@ -184,35 +175,50 @@ or in words: we have to change the Riesz representer of the (restricted/projecte
 # ╔═╡ 57cda07f-e432-46af-b771-5e5a3067feac
 md"""
 ## Example
-As an example we use the Rayleigh quotient ``f(x) = \frac{x^{\mathrm{T}}Ax}{x^{\mathrm{T}}x}`` for a given symmetric matrix ``A``.
+We continue with the Rayleigh Quotient from before, now just starting with the defintion of the Euclidean case in the embedding, the function ``F``.
 """
 
-# ╔═╡ e3b955b1-c780-4302-a605-190c3d10cd6f
-Random.seed!(42)
-
 # ╔═╡ c3f3aeba-2849-4715-94e2-0c44613a2ce9
-f̃(x) = x'*A*x/(x'*x);
+F(x) = x'*A*x/(x'*x);
 
 # ╔═╡ 786fce04-53ef-448d-9657-31208b35fb7e
[…]
