+++ pluto/AutomaticDifferentiation.jl
@@ -6,11 +6,11 @@ using InteractiveUtils
 
 # ╔═╡ b0769dfa-28cf-440e-9ba2-1ef488f171a9
 begin # dev mode – use global env on my machine
-    import Pkg
+    using Pkg: Pkg
     # careful: this is _not_ a reproducible environment
     # activate the global environment
     Pkg.activate()
-   using ReverseDiff, Manifolds, Manopt, Test, Random, LinearAlgebra, FiniteDiff
+    using ReverseDiff, Manifolds, Manopt, Test, Random, LinearAlgebra, FiniteDiff
 end
 
 # ╔═╡ 0213d26a-18ac-11ec-03fd-ada5992bcea8
@@ -20,12 +20,11 @@ md"""
 
 # ╔═╡ f3bc91ee-5871-4cba-ac89-190deb71ad0f
 md"""
-Since Manifolds.jl 0.6.7 the support of automatic differentiation support has been extended. With [ForwardDiff.jl](https://juliadiff.org/ForwardDiff.jl/stable/) or [ReverseDiff.jl](https://juliadiff.org/ReverseDiff.jl/) you can now use automatic differentiation on manifolds for embedded manifolds.
-"""
+Since [Manifolds.jl](https://juliamanifolds.github.io/Manifolds.jl/) 0.6.9 the support of automatic differentiation support has been extended.
 
-# ╔═╡ 35f02ab2-30d1-4c3c-8ba7-d07f391019e0
-md"""
-In general there is currenlty two ways to get the gradient of a function ``f\colon\mathcal M \to ℝ`` on a Riemannian manifold ``\mathcal M``.
+This tutorial explains how to use Euclidean tools to derive a gradient for a real-valued function ``F\colon \mathcal M → ℝ``. We will consider two methods: an intrinsic variant and a variant employing the embedding.
+
+While by default we use [FiniteDifferences.jl](https://juliadiff.org/FiniteDifferences.jl/latest/), you can also use [FiniteDiff.jl](https://github.com/JuliaDiff/FiniteDiff.jl), [ForwardDiff.jl](https://juliadiff.org/ForwardDiff.jl/stable/), [ReverseDiff.jl](https://juliadiff.org/ReverseDiff.jl/), or  [Zygote.jl](https://fluxml.ai/Zygote.jl/).
 """
 
 # ╔═╡ d9be6c2f-65fd-4685-9005-da22bf985e28
@@ -113,19 +112,19 @@ so we can compare it to the approximation by finite differences.
 
 # ╔═╡ 19747159-d383-4547-9315-0ed2494904a6
 begin
-	Random.seed!(42)
-	n = 200
-	A = randn(n+1,n+1)
-	A = Symmetric(A)
-	M = Sphere(n)
-	nothing
+    Random.seed!(42)
+    n = 200
+    A = randn(n + 1, n + 1)
+    A = Symmetric(A)
+    M = Sphere(n)
+    nothing
 end
 
 # ╔═╡ 41c204dd-6e4e-4a70-8f06-209a469e0680
-f1(p) = p'*A'p
+f1(p) = p' * A'p
 
 # ╔═╡ 2e33de5e-ffaa-422a-91d9-61f588ed1211
-gradf1(p) = 2*(A*p - p*p'*A*p) 
[…]
