### 2021-09-24 153f3b6
Fix a bug, where the iterate and gradient of a CG algorithm had to be of same type.
