+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Manopt"
 uuid = "0fc0a36d-df90-57f3-8f93-d78a9fc72bb5"
 authors = ["Ronny Bergmann <manopt@ronnybergmann.net>"]
-version = "0.3.12"
+version = "0.3.13"
 
 [deps]
 ColorSchemes = "35d6a980-a343-548e-a6ea-1d62b119f2f4"
@@ -24,7 +24,7 @@ ColorSchemes = "3.5.0"
 ColorTypes = "0.9.1, 0.10, 0.11"
 Colors = "0.11.2, 0.12"
 DataStructures = "0.17, 0.18"
-Manifolds = "0.5"
+Manifolds = "0.5, 0.6"
 ManifoldsBase = "0.12.3"
 Requires = "0.5, 1"
 StaticArrays = "0.12, 1.0"
