### 2021-10-06 6a2344c
Delete unneded (duplicate) structs, adapt to new Manifold name (AbstractManifold), runs formatter.
