+++ src/solvers/primal_dual_semismooth_Newton.jl
@@ -74,36 +74,45 @@ function primal_dual_semismooth_Newton(
     # variant=ismissing(Λ) ? :exact : :linearized,
     return_options=false,
     kwargs...,
-    ) where {
-    mT<:Manifold,
-    nT<:Manifold,
+) where {
+    mT<:AbstractManifold,
+    nT<:AbstractManifold,
     P,
     Q,
     T,
     RM<:AbstractRetractionMethod,
     IRM<:AbstractInverseRetractionMethod,
     VTM<:AbstractVectorTransportMethod,
-    }
+}
     p = PrimalDualSemismoothNewtonProblem(
-    M, N, cost, prox_F, diff_prox_F, prox_G_dual, diff_prox_G_dual, linearized_operator, adjoint_linearized_operator, Λ
+        M,
+        N,
+        cost,
+        prox_F,
+        diff_prox_F,
+        prox_G_dual,
+        diff_prox_G_dual,
+        linearized_operator,
+        adjoint_linearized_operator,
+        Λ,
     )
-    o =  PrimalDualSemismoothNewtonOptions(
-    m,
-    n,
-    x,
-    ξ,
-    primal_stepsize,
-    dual_stepsize;
-    # acceleration=acceleration,
-    # relaxation=relaxation,
-    stopping_criterion=stopping_criterion,
-    # relax=relax,
-    update_primal_base=update_primal_base, # TODO ?
-    update_dual_base=update_dual_base, # TODO ?
-    # variant=variant,
-    retraction_method=retraction_method,
-    inverse_retraction_method=inverse_retraction_method,
-    vector_transport_method=vector_transport_method,
+    o = PrimalDualSemismoothNewtonOptions(
+        m,
+        n,
+        x,
+        ξ,
+        primal_stepsize,
+        dual_stepsize;
+        # acceleration=acceleration,
+        # relaxation=relaxation,
+        stopping_criterion=stopping_criterion,
+        # relax=relax,
+        update_primal_base=update_primal_base, # TODO ?
+        update_dual_base=update_dual_base, # TODO ?
+        # variant=variant,
+        retraction_method=retraction_method,
+        inverse_retraction_method=inverse_retraction_method,
+        vector_transport_method=vector_transport_method,
     )
     o = decorate_options(o; kwargs...)
     resultO = solve(p, o)
@@ -114,9 +123,13 @@ function primal_dual_semismooth_Newton(
     end
 end
 
-function initialize_solver!(::PrimalDualSemismoothNewtonProblem, ::PrimalDualSemismoothNewtonOptions) end
+function initialize_solver!(
+    ::PrimalDualSemismoothNewtonProblem, ::PrimalDualSemismoothNewtonOptions
+) end
 
-function step_solver!(p::PrimalDualSemismoothNewtonProblem, o::PrimalDualSemismoothNewtonOptions, iter)
+function step_solver!(
+    p::PrimalDualSemismoothNewtonProblem, o::PrimalDualSemismoothNewtonOptions, iter
+)
     # do step
     primal_dual_step!(p, o)
[…]
