+++ src/plans/higher_order_primal_dual_plan.jl
@@ -1,7 +1,7 @@
 # TODO: change documentation
 # DONE: addapted struct and functions
 @doc raw"""
-    PrimalDualProblem {mT <: Manifold, nT <: Manifold} <: PrimalDualProblem} <: Problem
+    PrimalDualProblem {mT <: AbstractManifold, nT <: AbstractManifold} <: PrimalDualProblem} <: Problem
 
 Describes a Problem for the linearized Chambolle-Pock algorithm.
 
@@ -22,17 +22,19 @@ Describes a Problem for the linearized Chambolle-Pock algorithm.
     LinearizedPrimalDualProblem(M, N, cost, prox_F, prox_G_dual, forward_operator, adjoint_linearized_operator,Λ=forward_operator)
 
 """
-mutable struct PrimalDualSemismoothNewtonProblem{mT<:Manifold,nT<:Manifold} <: Problem
+mutable struct PrimalDualSemismoothNewtonProblem{
+    T,mT<:AbstractManifold,nT<:AbstractManifold
+} <: Problem{T}
     M::mT
     N::nT
     cost::Function
-    prox_F::Function
-    diff_prox_F::Function
-    prox_G_dual::Function
-    diff_prox_G_dual::Function
-    forward_operator::Function
-    adjoint_linearized_operator::Function
-    Λ::Union{Function,Missing}
+    prox_F!!::Function
+    diff_prox_F!!::Function
+    prox_G_dual!!::Function
+    diff_prox_G_dual!!::Function
+    forward_operator!!::Function
+    adjoint_linearized_operator!!::Function
+    Λ!!::Union{Function,Missing}
 end
 function PrimalDualSemismoothNewtonProblem(
     M::mT,
@@ -45,7 +47,7 @@ function PrimalDualSemismoothNewtonProblem(
     forward_operator,
     adjoint_linearized_operator,
     Λ=forward_operator,
-) where {mT<:Manifold,nT<:Manifold}
+) where {mT<:AbstractManifold,nT<:AbstractManifold}
     return PrimalDualSemismoothNewtonProblem{mT,nT}(
         M, N, cost, prox_F, prox_G_dual, forward_operator, adjoint_linearized_operator, Λ
     )
@@ -195,7 +197,13 @@ V_{x_k\gets m_k}\bigl(DΛ^*(m_k)\bigl[V_{n_k\gets n_{k-1}}ξ_{k-1} - ξ_k \bigr]
 ```
 where $V_{\cdot\gets\cdot}$ is the vector transport used in the [`ChambollePockOptions`](@ref)
 """
-function primal_residual(p::PrimalDualSemismoothNewtonProblem, o::PrimalDualSemismoothNewtonOptions, x_old, ξ_old, n_old)
+function primal_residual(
+    p::PrimalDualSemismoothNewtonProblem,
+    o::PrimalDualSemismoothNewtonOptions,
+    x_old,
+    ξ_old,
+    n_old,
+)
     return norm(
         p.M,
         o.x,
@@ -249,7 +257,13 @@ and for the `:exact` variant
 
 where in both cases $V_{\cdot\gets\cdot}$ is the vector transport used in the [`ChambollePockOptions`](@ref).
 """
-function dual_residual(p::PrimalDualSemismoothNewtonProblem, o::PrimalDualSemismoothNewtonOptions, x_old, ξ_old, n_old)
+function dual_residual(
+    p::PrimalDualSemismoothNewtonProblem,
+    o::PrimalDualSemismoothNewtonOptions,
+    x_old,
+    ξ_old,
+    n_old,
+)
     # if o.variant === :linearized (=> Always lineaized)
     return norm(
         p.N,
[…]
