### 2021-10-06 65a0f94
refactor to introduce all access functions for both mutating and nonmutating functions to the new solver.
