+++ src/plans/primal_dual_plan.jl
@@ -1,3 +1,5 @@
+abstract type AbstractPrimalDualProblem{T} <: Problem{T} end
+
 @doc raw"""
     PrimalDualProblem {T, mT <: Manifold, nT <: Manifold} <: PrimalDualProblem} <: Problem{T}
 
@@ -43,7 +45,7 @@ the second.
     > Journal of Mathematical Imaging and Vision 40(1), 120–145, 2011.
     > doi: [10.1007/s10851-010-0251-1](https://dx.doi.org/10.1007/s10851-010-0251-1)
 """
-mutable struct PrimalDualProblem{T,mT<:AbstractManifold,nT<:AbstractManifold} <: Problem{T}
+mutable struct PrimalDualProblem{T,mT<:AbstractManifold,nT<:AbstractManifold} <: AbstractPrimalDualProblem{T}
     M::mT
     N::nT
     cost::Function
@@ -88,19 +90,19 @@ Evaluate the proximal map of ``F`` stored within [`PrimalDualProblem`](@ref)
 
 which can also be computed in place of `y`.
 """
-get_primal_prox(::PrimalDualProblem, ::Any...)
+get_primal_prox(::AbstractPrimalDualProblem, ::Any...)
 
-function get_primal_prox(p::PrimalDualProblem{AllocatingEvaluation}, σ, x)
+function get_primal_prox(p::AbstractPrimalDualProblem{AllocatingEvaluation}, σ, x)
     return p.prox_F!!(p.M, σ, x)
 end
-function get_primal_prox(p::PrimalDualProblem{MutatingEvaluation}, σ, x)
+function get_primal_prox(p::AbstractPrimalDualProblem{MutatingEvaluation}, σ, x)
     y = allocate_result(p.M, get_primal_prox, x)
     return p.prox_F!!(p.M, y, σ, x)
 end
-function get_primal_prox!(p::PrimalDualProblem{AllocatingEvaluation}, y, σ, x)
+function get_primal_prox!(p::AbstractPrimalDualProblem{AllocatingEvaluation}, y, σ, x)
     return copyto!(p.M, y, p.prox_F!!(p.M, σ, x))
 end
-function get_primal_prox!(p::PrimalDualProblem{MutatingEvaluation}, y, σ, x)
+function get_primal_prox!(p::AbstractPrimalDualProblem{MutatingEvaluation}, y, σ, x)
     return p.prox_F!!(p.M, y, σ, x)
 end
 
@@ -116,19 +118,19 @@ Evaluate the proximal map of ``G_n^*`` stored within [`PrimalDualProblem`](@ref)
 
 which can also be computed in place of `y`.
 """
-get_dual_prox(::PrimalDualProblem, ::Any...)
+get_dual_prox(::AbstractPrimalDualProblem, ::Any...)
 
-function get_dual_prox(p::PrimalDualProblem{AllocatingEvaluation}, n, τ, ξ)
+function get_dual_prox(p::AbstractPrimalDualProblem{AllocatingEvaluation}, n, τ, ξ)
     return p.prox_G_dual!!(p.N, n, τ, ξ)
 end
-function get_dual_prox(p::PrimalDualProblem{MutatingEvaluation}, n, τ, ξ)
+function get_dual_prox(p::AbstractPrimalDualProblem{MutatingEvaluation}, n, τ, ξ)
     η = allocate_result(p.N, get_dual_prox, ξ)
     return p.prox_G_dual!!(p.N, η, n, τ, ξ)
 end
-function get_dual_prox!(p::PrimalDualProblem{AllocatingEvaluation}, η, n, τ, ξ)
[…]
