+++ pluto/AutomaticDifferentiation.jl
@@ -1,17 +1,14 @@
 ### A Pluto.jl notebook ###
-# v0.16.0
+# v0.16.1
 
 using Markdown
 using InteractiveUtils
 
+# ╔═╡ f88b15de-cec6-4bc8-9b68-2a407b5aeded
+using Manifolds, Manopt, Test, Random, LinearAlgebra
+
 # ╔═╡ b0769dfa-28cf-440e-9ba2-1ef488f171a9
-begin # dev mode – use global env on my machine
-    using Pkg: Pkg
-    # careful: this is _not_ a reproducible environment
-    # activate the global environment
-    Pkg.activate()
-    using ReverseDiff, Manifolds, Manopt, Test, Random, LinearAlgebra, FiniteDiff
-end
+using FiniteDiff, ReverseDiff
 
 # ╔═╡ 0213d26a-18ac-11ec-03fd-ada5992bcea8
 md"""
@@ -22,7 +19,7 @@ md"""
 md"""
 Since [Manifolds.jl](https://juliamanifolds.github.io/Manifolds.jl/) 0.6.9 the support of automatic differentiation support has been extended.
 
-This tutorial explains how to use Euclidean tools to derive a gradient for a real-valued function ``F\colon \mathcal M → ℝ``. We will consider two methods: an intrinsic variant and a variant employing the embedding.
+This tutorial explains how to use Euclidean tools to derive a gradient for a real-valued function ``F\colon \mathcal M → ℝ``. We will consider two methods: an intrinsic variant and a variant employing the embedding. These gradients can then be used within any gradient based optimisation algorithm in [Manopt.jl](https://manoptjl.org).
 
 While by default we use [FiniteDifferences.jl](https://juliadiff.org/FiniteDifferences.jl/latest/), you can also use [FiniteDiff.jl](https://github.com/JuliaDiff/FiniteDiff.jl), [ForwardDiff.jl](https://juliadiff.org/ForwardDiff.jl/stable/), [ReverseDiff.jl](https://juliadiff.org/ReverseDiff.jl/), or  [Zygote.jl](https://fluxml.ai/Zygote.jl/).
 """
@@ -37,9 +34,6 @@ In this Notebook we will take a look at a few possibilities to approximate or de
 
 Let's first load all packages we need.
 """
-
-# ╔═╡ f88b15de-cec6-4bc8-9b68-2a407b5aeded
-# using ReverseDiff, Manifolds, Manopt, Test, Random, LinearAlgebra
 
 # ╔═╡ 18d7459f-eed6-489b-a096-ac77ccd781af
 md"""
@@ -173,7 +167,7 @@ or in words: we have to change the Riesz representer of the (restricted/projecte
 
 # ╔═╡ 57cda07f-e432-46af-b771-5e5a3067feac
 md"""
-## Example
+### A continued Example
 We continue with the Rayleigh Quotient from before, now just starting with the defintion of the Euclidean case in the embedding, the function ``F``.
 """
 
[…]
