### 2021-10-12 cafabe3
Rephrase the SPD gradient known solution.
