+++ docs/make.jl
@@ -1,4 +1,4 @@
-using Manopt, Manifolds, Documenter, Literate
+using Manopt, Manifolds, Documenter, Literate, Pluto
 
 # generate examples using Literate
 tutorialsInputPath = joinpath(@__DIR__, "..", "src/tutorials")
@@ -33,10 +33,43 @@ for (i, tutorial) in enumerate(tutorials)
     push!(TutorialMenu, menuEntries[i] => joinpath(tutorialsRelativePath, tutorial * ".md"))
 end
 
+#
+#
+# Generate Pluto Tutorial HTMLs
+
+# First tutorial with AD
+pluto_src_folder = joinpath(@__DIR__, "..", "pluto/")
+pluto_output_folder = joinpath(@__DIR__, "src/", "pluto/")
+pluto_relative_path = "pluto/"
+mkpath(pluto_output_folder)
+#
+#
+# Please do not use the same name as for a(n old) literate Tutorial
+pluto_files = ["AutomaticDifferentiation",]
+pluto_titles = ["AD in Manopt",]
+for (i, f) in enumerate(pluto_files)
+    pluto_file = pluto_src_folder*f*".jl"
+    s = Pluto.ServerSession();
+    nb = Pluto.SessionActions.open(s, pluto_file; run_async=false)
+    html_contents = Pluto.generate_html(nb; binder_url_js="undefined")
+    open(pluto_output_folder*f*"_export.html","w") do io
+        println(io,html_contents)
+    end
+    open(pluto_output_folder*f*".md","w") do io
+        println(io,"""
+        ```@raw html
+        <iframe style="border:none; width:100%;" sandbox="allow-same-origin allow-scripts" frameborder="0" scrolling="no"
+        onload="this.style.height=(this.contentWindow.document.body.scrollHeight+20)+'px';" src="$(f)_export.html"></iframe>
+        ```
+        """
+        )
+    end
+    push!(TutorialMenu, pluto_titles[i] => joinpath(pluto_relative_path, f * ".md"))
+end
+
 generated_path = joinpath(@__DIR__, "src")
 base_url = "https://github.com/JuliaManifolds/Manopt.jl/blob/master/"
 isdir(generated_path) || mkdir(generated_path)
-
 open(joinpath(generated_path, "contributing.md"), "w") do io
     # Point to source license file
     println(
