### 2021-10-13 64907da
Switch to a (though hidden) local environment in the Pluto Notebook, switch to a static height of the iframe.
