+++ pluto/AutomaticDifferentiation.jl
@@ -5,7 +5,20 @@ using Markdown
 using InteractiveUtils
 
 # ╔═╡ f88b15de-cec6-4bc8-9b68-2a407b5aeded
-using Manifolds, Manopt, Test, Random, LinearAlgebra
+begin
+	using Pkg: Pkg
+    Pkg.activate(mktempdir())
+    Pkg.add([
+        Pkg.PackageSpec(; name="Manifolds", version="0.7"),
+        Pkg.PackageSpec(; name="Manopt", version="0.3.13"),
+        Pkg.PackageSpec(; name="FiniteDiff", version="2.8.1"),
+        Pkg.PackageSpec(; name="ReverseDiff", version="1.9.0"),
+#        Pkg.PackageSpec(; name="PlutoUI"),
+     ])
+end
+
+# ╔═╡ 856f336c-e232-4f1f-b1ac-759b4558acd1
+using Manifolds, Manopt, Random, LinearAlgebra
 
 # ╔═╡ b0769dfa-28cf-440e-9ba2-1ef488f171a9
 using FiniteDiff, ReverseDiff
@@ -285,530 +298,12 @@ md"""
 This tutorial illustrates how to use tools from Euclidean spaces, finite differences or automatic differentiation, to compute gradients on Riemannian manifolds. The scheme allows to use _any_ differentiation framework within the embedding to derive a Riemannian gradient.
 """
 
-# ╔═╡ 00000000-0000-0000-0000-000000000001
-PLUTO_PROJECT_TOML_CONTENTS = """
-[deps]
-FiniteDiff = "6a86dc24-6348-571c-b903-95158fe2bd41"
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Manifolds = "1cead3c2-87b3-11e9-0ccd-23c62b72b94e"
-Manopt = "0fc0a36d-df90-57f3-8f93-d78a9fc72bb5"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[compat]
-FiniteDiff = "~2.8.1"
-Manifolds = "~0.7.0"
-Manopt = "~0.3.13"
-ReverseDiff = "~1.9.0"
-"""
-
-# ╔═╡ 00000000-0000-0000-0000-000000000002
-PLUTO_MANIFEST_TOML_CONTENTS = """
-# This file is machine-generated - editing it directly is not advised
-
-[[AbstractFFTs]]
-deps = ["LinearAlgebra"]
-git-tree-sha1 = "485ee0867925449198280d4af84bdb46a2a404d0"
-uuid = "621f4979-c628-5d54-868e-fcf4e3e8185c"
-version = "1.0.1"
-
-[[ArgTools]]
-uuid = "0dad84c5-d112-42e6-8d28-ef12dabb789f"
-
-[[ArnoldiMethod]]
-deps = ["LinearAlgebra", "Random", "StaticArrays"]
-git-tree-sha1 = "f87e559f87a45bece9c9ed97458d3afe98b1ebb9"
-uuid = "ec485272-7323-5ecc-a04f-4719b315124d"
-version = "0.1.0"
-
-[[ArrayInterface]]
[…]
