### 2021-10-13 ed80268
bump version and be a little less strict on versions in Pluto.
