### 2021-10-13 e6a1aaa
A tutorial on AD (and FD) on manifolds and in the embedding in Pluto (#94)
* Start a notebook on AD.
* include Pluto Rendering as iframe in the docs.

Co-authored-by: Mateusz Baran <mateuszbaran89@gmail.com>
