+++ pluto/AutomaticDifferentiation.jl
@@ -1,0 +1,347 @@
+### A Pluto.jl notebook ###
+# v0.16.1
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ f88b15de-cec6-4bc8-9b68-2a407b5aeded
+begin
+    using Pkg: Pkg
+    Pkg.activate(mktempdir())
+    Pkg.add([
+        Pkg.PackageSpec(; name="Manifolds", version="0.7"),
+        Pkg.PackageSpec(; name="Manopt", version="0.3"),
+        Pkg.PackageSpec(; name="FiniteDiff", version="2.8"),
+        Pkg.PackageSpec(; name="ReverseDiff", version="1.9"),
+        # Pkg.PackageSpec(; name="PlutoUI"),
+    ])
+end
+
+# ╔═╡ 856f336c-e232-4f1f-b1ac-759b4558acd1
+using Manifolds, Manopt, Random, LinearAlgebra
+
+# ╔═╡ b0769dfa-28cf-440e-9ba2-1ef488f171a9
+using FiniteDiff, ReverseDiff
+
+# ╔═╡ 0213d26a-18ac-11ec-03fd-ada5992bcea8
+md"""
+# Using (Euclidean) AD in Manopt.jl
+"""
+
+# ╔═╡ f3bc91ee-5871-4cba-ac89-190deb71ad0f
+md"""
+Since [Manifolds.jl](https://juliamanifolds.github.io/Manifolds.jl/latest/) 0.7 the support of automatic differentiation support has been extended.
+
+This tutorial explains how to use Euclidean tools to derive a gradient for a real-valued function ``F\colon \mathcal M → ℝ``. We will consider two methods: an intrinsic variant and a variant employing the embedding. These gradients can then be used within any gradient based optimisation algorithm in [Manopt.jl](https://manoptjl.org).
+
+While by default we use [FiniteDifferences.jl](https://juliadiff.org/FiniteDifferences.jl/latest/), you can also use [FiniteDiff.jl](https://github.com/JuliaDiff/FiniteDiff.jl), [ForwardDiff.jl](https://juliadiff.org/ForwardDiff.jl/stable/), [ReverseDiff.jl](https://juliadiff.org/ReverseDiff.jl/), or  [Zygote.jl](https://fluxml.ai/Zygote.jl/).
+"""
+
+# ╔═╡ d9be6c2f-65fd-4685-9005-da22bf985e28
+md"""
+In this Notebook we will take a look at a few possibilities to approximate or derive the gradient of a function ``f:\mathcal M \to ℝ`` on a Riemannian manifold, without computing it yourself. There is mainly two different philosophies:
+
+1. Working _instrinsically_, i.e. stay on the manifold and in the tangent spaces. Here, we will consider approximating the gradient by forward differences.
+
+2. Working in an embedding – there we can use all tools from functions on Euclidean spaces – finite differences or automatic differenciation – and then compute the corresponding Riemannian gradient from there.
+
+Let's first load all packages we need.
+"""
+
+# ╔═╡ 18d7459f-eed6-489b-a096-ac77ccd781af
+md"""
+## 1. (Intrinsic) Forward Differences
+
[…]
