### 2021-10-15 3c35f6c
begins to implement the Riemannian Augmented Lagrangian Method
