+++ docs/src/solvers/Riemannian_augmented_Lagrangian_method.md
@@ -1,0 +1,18 @@
+# [Riemannian augmented Lagrangian method](@id AugmentedLagrangianSolver)
+
+```@meta
+CurrentModule = Manopt
+```
+
+```@docs
+  Riemannian_augmented_Lagrangian_method
+  Riemannian_augmented_Lagrangian_method!
+```
+
+## Options
+
+```@docs
+RALMOptions
+```
+
+## Literature
