### 2021-10-21 c9731bd
adds function constraints_detail
