### 2021-10-24 dad8e6d
renames Riemannian augmented Lagrangian method to just augmented Lagrangian method, changes the options accordingly and keeps working on both
