+++ src/solvers/Riemannian_augmented_Lagrangian_method.jl
@@ -1,175 +1,0 @@
-# Recreating Changshuo Liu's Matlab source code in Julia
-# original code by Changshuo Liu: https://github.com/losangle/Optimization-on-manifolds-with-extra-constraints/blob/master/solvers/almbddmultiplier.m
-# uses Riemannian limited memory BFGS solver
-# get n_ineq_constraint, n_eq_constraint
-
-function Riemannian_augmented_Lagrangian_method(
-    M::Manifold,
-    F::TF,
-    n_ineq_constraint::Int,
-    n_eq_constraint::Int;
-    x::random_point(M),
-    kwargs...,
-) where {TF}
-    x_res = allocate(x)
-    copyto!(Ref(M), x_res, x)
-    return particle_swarm!(M, F, n_ineq_constraint, n_eq_constraint; x=x_res, kwargs...)
-end
-
-function Riemannian_augmented_Lagrangian_method!(
-    M::Manifold,
-    F::TF,
-    n_ineq_constraint::Int,
-    n_eq_constraint::Int;
-    x::random_point(M),
-    max_inner_iter::Int=200,
-    num_outer_itertgn::Int=30,
-    ϵ::Real=1e-3, #(starting)tolgradnorm
-    ϵ_min::Real=1e-6, #endingtolgradnorm
-    bound::Int=20, ###why not Real?
-    λ::Vector=ones(n_ineq_constraint,1),
-    γ::Vector=ones(n_eq_constraint,1),
-    ρ::Real=1.0, ###why int in Matlab code?
-    τ::Real=0.8,
-    θ_ρ::Real=0.3, 
-    θ_ϵ::Real=(ϵ_min/ϵ)^(1/num_outer_itertgn), ###this does not need to be a parameter, just defined somewhere
-    oldacc::Real=Inf, ###this does not need to be a parameter, just defined somewhere
-    min_stepsize::Real= ### find realistic value, put in stopping criterion alongside the ϵ condition
-    stopping_criterion::StoppingCriterion==StopWhenAny(StopAfterIteration(300), StopWhenAll()), #maxOuterIter
-    kwargs...,
-) where {TF}
-    p = CostProblem(M, F, n_ineq_constraint, n_eq_constraint)
-    o = RALMOptions(
-        x,
-        max_inner_iter,
-        num_outer_itertgn,
-        ϵ,
-        ϵ_min,
-        bound,
-        λ,
-        γ,
-        ρ,
-        τ,
-        θ_ρ,
-        θ_ϵ,
-        oldacc,
-        stopping_criterion,
-    )
-    o = decorate_options(o; kwargs...)
-    resultO = solve(p, o)
-    if return_options
-        return resultO
-    else
-        return get_solver_result(resultO)
-    end
-end
-
-#
-# Solver functions
-#
-function initialize_solver!(p::CostProblem, o::RALMOptions)
-###
-end
-function step_solver!(p::CostProblem, o::RALMOptions, iter)
-    # use subsolver(Riemannian limited memory BFGS) to minimize the augmented Lagrangian within a tolerance ϵ and with max_inner_iter
-    
-    ###o.x=
-
-    # update multipliers
-    newacc=0
-    for i in 1:o.n_ineq_constraint
-        ### how can I access the cost of the inequality (cost_iter)?
-        newacc = max(newacc, abs(max(-λ[i]/ρ, cost_iter)))
-        λ[i]=min(o.bound, max(λ[i] + o.ρ * cost_iter, 0))
-    end 
-    for j in 1:o.n_eq_constraint
-        ### how can I access the cost of the equality (cost_iter)?
[…]
