### 2021-10-31 e173ccf
adds a new stopping criterion and uses it for the ALM
