### 2021-11-03 a10ded9
Precompute the inner-product of the gradient with the search direction one, instead of computing it several times for each linesearch backtracking.
