### 2021-11-03 JuliaManifolds/Manopt.jl#96 @sweichwald
Minor comments on joss/paper.md
Comments on software paper as per 6f74fee as part of https://github.com/openjournals/joss-reviews/issues/3866#event-5537686057.

---

Only minor comments:
- 10: "Nelder Mead" -> "Nelder-Mead"
- 11: "quasi Newton" -> "Quasi-Newton"
- 41: possibly missing an "and" in "framework can"
- 55: missing comma after "manifold"
- 57: maybe mention an example of a manifold with no closed forms in parentheses
- 57: "more generally arbitrary retractions" -> "more general arbitrary retractions"
- 59: maybe rephrase; it seems there are two reasons for the use of retractions: closed form exponential maps are a) unavailable, or b) costly to compute; thus, it may be odd to state that these approximate algorithms are "more efficient" (than what?) especially in cases when, as stated in the preceding sentence, the closed form exponential map may not even be available
- 72ff: it may be helpful to complement each item by mentioning the corresponding names in Manopt.jl explicitly, e.g. one could mention `conjugate_gradient_descent` and `coefficient=DirectionUpdateRule` (`SteepestDirectionUpdateRule`, `ConjugateDescentCoefficient`, `DaiYuanCoefficient`, `FletcherReevesCoefficient`, `HagerZhangCoefficient`, `HeestenesStiefelCoefficient`, `LiuStoreyCoefficient`, and `PolakRibiereCoefficient`) somewhat mimicking the nice layout of the online documentation and providing readers already a more concrete glimpse into Manopt.jl's scope and potential
- 93: typo, redundant "the"
- 95: for readers less familiar with optimization on manifolds, it may be more instructive to drastically reduce `n` and show the Karcher mean in contrast to the Euclidean mean so as to highlight that the Karcher mean still sits on the sphere (e.g. "a meeting half-way between Trondheim and Lausanne would take place in and not under central Europe" :-))?
- 109ff: can the print output precision be meaningfully adjusted or aligned (the latter would, for example, help to visually identify the changes in `F(x)`)
- 122: "There are" -> "The"
- 144: "Bačák, M. (2014)" appears to be a duplicate of 142
