### 2021-11-03 7e9334f
Adress points from the review of @sweichwald, part I.
