+++ joss/paper.md
@@ -20,7 +20,7 @@ bibliography: bibliography.bib
 # Summary
 
 [`Manopt.jl`](https://manoptjl.org) provides a set of optimization algorithms for optimization problems given on a Riemannian manifold $\mathcal M$.
-Based on a generic optimization framework together with the interface `ManifoldsBase.jl` for Riemannian manifolds, classical and recently developed methods are provided in an efficient implementation. Algorithms include the derivative free Particle Swarm and Nelder Mead algorithms as well as a classical gradient, conjugate gradient and stochastic gradient descent. Furthermore, quasi Newton methods like a Riemannian L-BFGS [@HuangGallivanAbsil:2015:1] and nonsmooth optimization algorithms like a Cyclic Proximal Point Algorithm [@Bacak:2014:1], a (parallel) Douglas-Rachford algorithm [@BergmannPerschSteidl:2016:1] and a Chambolle-Pock algorithm [@BergmannHerzogSilvaLouzeiroTenbrinckVidalNunez:2021:1] are provided together with several basic cost functions, gradients and proximal maps as well as debug and record capabilities.
+Based on a generic optimization framework together with the interface `ManifoldsBase.jl` for Riemannian manifolds, classical and recently developed methods are provided in an efficient implementation. Algorithms include the derivative free Particle Swarm and Nelder–Mead algorithms as well as a classical gradient, conjugate gradient and stochastic gradient descent. Furthermore, quasi-Newton methods like a Riemannian L-BFGS [@HuangGallivanAbsil:2015:1] and nonsmooth optimization algorithms like a Cyclic Proximal Point Algorithm [@Bacak:2014:1], a (parallel) Douglas-Rachford algorithm [@BergmannPerschSteidl:2016:1] and a Chambolle-Pock algorithm [@BergmannHerzogSilvaLouzeiroTenbrinckVidalNunez:2021:1] are provided together with several basic cost functions, gradients and proximal maps as well as debug and record capabilities.
 
 # Statement of Need
 
@@ -41,7 +41,7 @@ Further examples and a thorough introduction can be found in [@AbsilMahonySepulc
 
 For a user facing an optimization problem on a manifold, there are two obstacles to the actual numerical optimization: on the one hand, a suitable implementation of the manifold at hand is required, for example how to evaluate the above mentioned geodesics. On the other hand, an implementation of the optimization algorithm that employs said methods from the manifold, such that the algorithm can be applied to the cost function $f$ a user already has.
 
-Using the interface for manifolds, `ManifoldsBase.jl`, the algorithms are implemented in the optimization framework can therefore be used with any manifold from `Manifolds.jl` [@AxenBaranBergmannRzecki:2021:1], a library of efficiently implemented Riemannian manifolds.
+Using the interface for manifolds, `ManifoldsBase.jl`, the algorithms are implemented in the optimization framework. They can therefore be used with any manifold from `Manifolds.jl` [@AxenBaranBergmannRzecki:2021:1], a library of efficiently implemented Riemannian manifolds.
 `Manopt.jl` provides a low level entry to optimization on manifolds while also providing efficient implementations, that can easily be extended to cover own manifolds.
 
 # Functionality
@@ -51,7 +51,9 @@ The framework includes a generic way to specify a step size and a stopping crite
 Each of the algorithms has a high level interface to make it easy to use the algorithms directly.
 
 An optimization task in `Manopt.jl` consists of a `Problem p` and `Options o`,
[…]
