+++ joss/bibliography.bib
@@ -116,17 +116,6 @@
     URL     = {https://www.manopt.org}
 }
 
-@article{Bacak:2014:1,
-    AUTHOR       = {Bačák, M.},
-    DATE         = {2014},
-    DOI          = {10.1137/140953393},
-    JOURNALTITLE = {SIAM Journal on Optimization},
-    NUMBER       = {3},
-    PAGES        = {1542--1566},
-    TITLE        = {Computing medians and means in Hadamard spaces},
-    VOLUME       = {24}
-}
-
 @incollection{BorckmansIshtevaAbsil2010,
     DOI       = {10.1007/978-3-642-15461-4_2},
     YEAR      = 2010,
@@ -181,4 +170,16 @@
     PAGES        = {509--541},
     TITLE        = {Riemannian center of mass and mollifier smoothing},
     VOLUME       = {30}
+}
+
+@proceedings{BendokatZimmermann:2021,
+    AUTHOR    = {Bendokat, T. and Zimmermann, R.},
+    DATE      = {2021},
+    DOI       = {10.1007/978-3-030-80209-7_82},
+    PAGES     = {763--771},
+    PUBLISHER = {Springer International Publishing},
+    TITLE     = {Efficient Quasi-Geodesics on the Stiefel Manifold},
+    URL       = {https://doi.org/10.1007%2F978-3-030-80209-7_82},
+    EDITOR    = {Nielsen F., Barbaresco F.},
+    BOOKTITLE = {Geometric Science of Information. GSI 2021}
 }
