### 2021-11-04 JuliaManifolds/Manopt.jl#96 @sweichwald
Re 109ff: Opening an issue to keep track of future development of this feature is certainly ok!
