### 2021-11-04 83f4af1
adds an example for a manifold without a closed form expression for exp.
