+++ joss/paper.md
@@ -51,7 +51,7 @@ The framework includes a generic way to specify a step size and a stopping crite
 Each of the algorithms has a high level interface to make it easy to use the algorithms directly.
 
 An optimization task in `Manopt.jl` consists of a `Problem p` and `Options o`,
-The `Problem` consists of all static information like the cost function and a potential gradient of the optimization task. The `Options` specify the type of algorithm and the settings and data required to run the algorithm. For example by default most options specify that the exponential map, which generalizes the notion of addition to the manifold, should be used and the algorithm steps are performed following an acceleration free curve on the manifold. This might not be known in closed form for some manifolds and hence also more general arbitrary retractions can be specified for this instead.
+The `Problem` consists of all static information like the cost function and a potential gradient of the optimization task. The `Options` specify the type of algorithm and the settings and data required to run the algorithm. For example by default most options specify that the exponential map, which generalizes the notion of addition to the manifold, should be used and the algorithm steps are performed following an acceleration free curve on the manifold. This might not be known in closed form for some manifolds, e.g. the [`Spectrahedron`](https://juliamanifolds.github.io/Manifolds.jl/v0.7/) does not have -- to the best of the authors knowledge -- a closed form expression for the exponential map. Hence also more general arbitrary retractions can be specified for this instead.
 Retractions are first order approximations for the exponential map. They provide an alternative to the acceleration free form, if no closed form solution is known. Otherwise, a retraction might also be chosen, when their evaluation is computationally cheaper than to use the exponential map, especially if their approximation error can be stated, see e.g. [@BendokatZimmermann:2021].
 
 Similarly, tangent vectors at different points are identified by a vector transport, which by default is the parallel transport.
