### 2021-11-04 d74d5d2
Add links to solvers and fixes a few typos in the docs.
