+++ joss/paper.md
@@ -63,18 +63,19 @@ Using a decorator pattern, the `Options` can be encapsulated in `DebugOptions` a
 
 In the current version `Manopt.jl` version 0.3.12 the following algorithms are available
 
-* Alternating Gradient Descent
-* Chambolle-Pock [@BergmannHerzogSilvaLouzeiroTenbrinckVidalNunez:2021:1]
-* Conjugate Gradient Descent, including eight update rules
-* Cyclic Proximal Point [@Bacak:2014:1]
-* (parallel) Douglas–Rachford [@BergmannPerschSteidl:2016:1]
-* Gradient Descent, including direction update rules including Momentum, Average, and a Nestorv-type one
-* Nelder-Mead
-* Particle Swarm Optimization [@BorckmansIshtevaAbsil2010]
-* Quasi-Newton, with the BFGS, DFP, Broyden and a symmetric rank 1 update, their inverse updates as well as a limited memory variant of (inverse) BFGS [@HuangGallivanAbsil:2015:1]
-* Stochastic Gradient Descent
-* Subgradient Method
-* Trust Regions, with inner Steihaug-Toint TCG solver [@AbsilBakerGallivan2006]
+* Alternating Gradient Descent ([`alternating_gradient_descent](https://manoptjl.org/v0.3/solvers/alternating_gradient_descent.html))
+* Chambolle-Pock ([`ChambollePock`](https://manoptjl.org/v0.3/solvers/ChambollePock.html)) [@BergmannHerzogSilvaLouzeiroTenbrinckVidalNunez:2021:1]
+* Conjugate Gradient Descent ([`conjugate_gradient_descent`](https://manoptjl.org/v0.3/solvers/conjugate_gradient_descent.html)), which includes eight direction update rules using the `coefficient` keyword:
+  [`SteepestDirectionUpdateRule`](https://manoptjl.org/v0.3/solvers/conjugate_gradient_descent.html#Manopt.SteepestDirectionUpdateRule),   [`ConjugateDescentCoefficient`](https://manoptjl.org/v0.3/solvers/conjugate_gradient_descent.html#Manopt.ConjugateDescentCoefficient). [`DaiYuanCoefficient`](https://manoptjl.org/v0.3/solvers/conjugate_gradient_descent.html#Manopt.DaiYuanCoefficient), [`FletcherReevesCoefficient`](https://manoptjl.org/v0.3/solvers/conjugate_gradient_descent.html#Manopt.FletcherReevesCoefficient), [`HagerZhangCoefficient`](https://manoptjl.org/v0.3/solvers/conjugate_gradient_descent.html#Manopt.HagerZhangCoefficient), [`HeestenesStiefelCoefficient`](https://manoptjl.org/v0.3/solvers/conjugate_gradient_descent.html#Manopt.HeestenesStiefelCoefficient), [`LiuStoreyCoefficient`](https://manoptjl.org/v0.3/solvers/conjugate_gradient_descent.html#Manopt.LiuStoreyCoefficient), and[`PolakRibiereCoefficient`](https://manoptjl.org/v0.3/solvers/conjugate_gradient_descent.html#Manopt.PolakRibiereCoefficient)
+* Cyclic Proximal Point ([`cyclic_proximal_point`](https://manoptjl.org/v0.3/solvers/cyclic_proximal_point.html)) [@Bacak:2014:1]
+* (parallel) Douglas–Rachford ([`DouglasRachford`](https://manoptjl.org/v0.3/solvers/DouglasRachford.html)) [@BergmannPerschSteidl:2016:1]
[…]
