### 2021-11-04 cd1b038
spread the points in the experiment, compare to Euclidean mean.
