+++ joss/paper.md
@@ -80,19 +80,22 @@ In the current version `Manopt.jl` version 0.3.12 the following algorithms are a
 # Example
 
 `Manopt.jl` is registered in the general Julia registry and can hence be installed typing `]add Manopt` in Julia REPL.
-Given the `Sphere` from `Manifolds.jl` and a set of unit vectors $p_1,...,p_N\in\mathbb R^3$, where $N$ is the number of data points.
-we can compute the generalization of the mean, called the Riemannian Center of Mass [@Karcher:1977:1], which is defined as the minimizer of the squared distances to the given data
+Given the [`Sphere`](https://juliamanifolds.github.io/Manifolds.jl/v0.7/manifolds/sphere.html) from `Manifolds.jl` and a set of unit vectors $p_1,...,p_N\in\mathbb R^3$, where $N$ is the number of data points.
+we can compute the generalization of the mean, called the Riemannian Center of Mass [@Karcher:1977:1], which is defined as the minimizer of the squared distances to the given datan – a property the mean in vector spaces fulfills –
 
 $$ \operatorname*{arg\,min}_{x\in\mathcal M}\quad \displaystyle\sum_{k=1}^Nd_{\mathcal M}(x, p_k)^2, $$
 
-where $d_{\mathcal M}$ denotes length of a shortest geodesic connecting the two points in the arguments. It is called the Riemannian distance. For the sphere this distance is given by the length of the shorter great arc connecting the two points.
+where $d_{\mathcal M}$ denotes length of a shortest geodesic connecting the two points in the arguments.
+It is called the Riemannian distance. For the sphere this [`distance`](https://juliamanifolds.github.io/Manifolds.jl/v0.7/manifolds/sphere.html#ManifoldsBase.distance-Tuple{AbstractSphere,%20Any,%20Any}) is given by the length of the shorter great arc connecting the two points.
 
 ```julia
 using Manopt, Manifolds, LinearAlgebra, Random
 Random.seed!(42)
 M = Sphere(2)
-n = 100
-pts = [ normalize(rand(3)) for _ in 1:n ]
+n = 40
+p = 1/sqrt(3) .* ones(3)
+B = DefaultOrthonormalBasis()
+pts = [ exp(M, p, get_vector(M, p, π/8*randn(2), B)) for _ in 1:n ]
 
 F(M, y) = sum(1/(2*n) * distance.(Ref(M), pts, Ref(y)).^2)
 gradF(M, y) = sum(1/n * grad_distance.(Ref(M), pts, Ref(y)))
@@ -100,9 +103,13 @@ gradF(M, y) = sum(1/n * grad_distance.(Ref(M), pts, Ref(y)))
 x_mean = gradient_descent(M, F, gradF, pts[1])
 ```
 
-Both the data `pts` and the resulting mean are shown in the following figure.
+The resulting `x_mean` minimizes the (Riemannian) distances squared but is especially a point of unit norm.
+Compared to `mean(pts)`, which computes the mean in the embedding of the sphere, the $\mathbb R^3$, yields a point “inside” the sphere,
+since its norm is approximately `0.858`. But even projecting this back onto the sphere, yields a point that does not fulfill the property of minimizing the squared distances.
 
-![100 random points `pts` and the result from the gradient descent to compute the `x_mean` (orange).](src/img/MeanIllustr.png)
+In the following figure the data `pts` (teal) and the resulting mean (orange) as well as the projected Euclidean mean (small, cyan) are shown.
+
+![40 random points `pts` and the result from the gradient descent to compute the `x_mean` (orange) compared to a projection of their (Eucliean) mean onto the sphere (cyan).](src/img/MeanIllustr.png)
 
[…]
