### 2021-11-04 8a39c67
switch to a non-utf8-sigma for randn.
