### 2021-11-04 JuliaManifolds/Manopt.jl#96 @kellertuer
* I adressed line 57 (manifold without closed from exp) in 83f4af1
* I added the remark on l-72ff – the links to the solvers – in d74d5d2, where the links all point to the most recent version of 0.3.x (just in case one of them breaks in a future version due to restructure of docs/package). Note that the joss branch – when merged – will also trigger a new version (0.3.15 as of now) adding the paper to master ans fixing a few typos I noticed along the was
* Concerning the example (l. 95) I reduced the number of points to 40 and introduced an explanation/comparison to the (Euclidean) mean in R3 and the differences to it. See 8a39c67 for the main part ( 8a39c67 introduces a non-utf8 sigma instead of `π/8`, since the Pandoc OpenJOSS style does not support π in source code blocks).
* thanks for the feedback on the Debug, that is now tracked (including a modelling) in #97.
