### 2021-11-04 759105c
Fixes two typos.
