### 2021-11-08 cb43159
renames the new stopping criterion and uses symbol
