### 2021-11-08 33ab999
adds new functions to the constrained plan
