### 2021-11-08 790906a
adapts name change of stopping criterion, turns get_Lagrangian_cost and get_Lagrangian_gradient into get_Lagrangian_cost_function and get_Lagrangian_gradient_function and lets them return functions instead of values
