### 2021-11-08 796b4d7
does some debugging
