+++ src/plans/plan.jl
@@ -6,6 +6,7 @@ include("options.jl")
 include("debug_options.jl")
 include("record_options.jl")
 
+include("alm_plan.jl")
 include("cost_plan.jl")
 include("gradient_plan.jl")
 include("conjugate_gradient_plan.jl")
