### 2021-11-09 a0d27cf
Followed adjustments from 'JuliaFormatter.jl'.
