+++ src/plans/stepsize.jl
@@ -50,7 +50,7 @@ mutable struct DecreasingStepsize <: Stepsize
     factor::Float64
     subtrahend::Float64
     exponent::Float64
-    function DecreasingStepsize(l::Real = 1.0, f::Real = 1.0, a::Real = 0.0, e::Real = 1.0)
+    function DecreasingStepsize(l::Real=1.0, f::Real=1.0, a::Real=0.0, e::Real=1.0)
         return new(l, f, a, e)
     end
 end
@@ -109,19 +109,16 @@ mutable struct ArmijoLinesearch{TRM<:AbstractRetractionMethod} <: Linesearch
     sufficientDecrease::Float64
     stepsizeOld::Float64
     function ArmijoLinesearch(
-        s::Float64 = 1.0,
-        r::AbstractRetractionMethod = ExponentialRetraction(),
-        contractionFactor::Float64 = 0.95,
-        sufficientDecrease::Float64 = 0.1,
+        s::Float64=1.0,
+        r::AbstractRetractionMethod=ExponentialRetraction(),
+        contractionFactor::Float64=0.95,
+        sufficientDecrease::Float64=0.1,
     )
         return new{typeof(r)}(s, r, contractionFactor, sufficientDecrease, s)
     end
 end
 function (a::ArmijoLinesearch)(
-    p::GradientProblem,
-    o::Options,
-    ::Int,
-    η = -get_gradient(p, o.x),
+    p::GradientProblem, o::Options, ::Int, η=-get_gradient(p, o.x)
 )
     a.stepsizeOld = linesearch_backtrack(
         p.M,
@@ -161,9 +158,9 @@ function linesearch_backtrack(
     s,
     decrease,
     contract,
-    retr::AbstractRetractionMethod = ExponentialRetraction(),
-    η::T = -gradFx,
-    f0 = F(x),
+    retr::AbstractRetractionMethod=ExponentialRetraction(),
+    η::T=-gradFx,
+    f0=F(x),
 ) where {TF,T}
     xNew = retract(M, x, s * η, retr)
     fNew = F(xNew)
@@ -264,9 +261,7 @@ with the Fields above in their order as optional arguments.
 This method returns the functor to perform nonmonotone line search.
 """
 mutable struct NonmonotoneLinesearch{
-    TRM<:AbstractRetractionMethod,
-    VTM<:AbstractVectorTransportMethod,
-    T<:AbstractVector,
+    TRM<:AbstractRetractionMethod,VTM<:AbstractVectorTransportMethod,T<:AbstractVector
 } <: Linesearch
     retraction_method::TRM
     vector_transport_method::VTM
@@ -279,16 +274,16 @@ mutable struct NonmonotoneLinesearch{
     strategy::Symbol
     storage::StoreOptionsAction
     function NonmonotoneLinesearch(
-        initial_stepsize::Float64 = 1.0,
-        retraction_method::AbstractRetractionMethod = ExponentialRetraction(),
-        vector_transport_method::AbstractVectorTransportMethod = ParallelTransport(),
-        stepsize_reduction::Float64 = 0.5,
-        sufficient_decrease::Float64 = 1e-4,
-        memory_size::Int = 10,
-        min_stepsize::Float64 = 1e-3,
-        max_stepsize::Float64 = 1e3,
-        strategy::Symbol = :direct,
-        storage::StoreOptionsAction = StoreOptionsAction((:x, :gradient)),
+        initial_stepsize::Float64=1.0,
+        retraction_method::AbstractRetractionMethod=ExponentialRetraction(),
+        vector_transport_method::AbstractVectorTransportMethod=ParallelTransport(),
+        stepsize_reduction::Float64=0.5,
[…]
