### 2021-11-11 aa320e9
adapts documentation of stopping criterion to the name change
