+++ docs/src/solvers/augmented_Lagrangian_method.md
@@ -1,0 +1,18 @@
+# [Augmented Lagrangian method](@id AugmentedLagrangianSolver)
+
+```@meta
+CurrentModule = Manopt
+```
+
+```@docs
+  augmented_Lagrangian_method
+  augmented_Lagrangian_method!
+```
+
+## Options
+
+```@docs
+ALMOptions
+```
+
+## Literature
