### 2021-11-11 96ce673
fixes some typos and marks not compiling block of code for discussion
