### 2021-11-11 9d2df56
adds the algorithm and the other files needed for it into the Manopt framework
