+++ docs/make.jl
@@ -65,6 +65,7 @@ makedocs(;
         "Solvers" => [
             "Introduction" => "solvers/index.md",
             "Alternating Gradient Descent" => "solvers/alternating_gradient_descent.md",
+            "Augmented Lagrangian method" => "solvers/augmented_Lagrangian_method.md",
             "Chambolle-Pock" => "solvers/ChambollePock.md",
             "Conjugate gradient descent" => "solvers/conjugate_gradient_descent.md",
             "Cyclic Proximal Point" => "solvers/cyclic_proximal_point.md",
@@ -73,7 +74,6 @@ makedocs(;
             "Nelder–Mead" => "solvers/NelderMead.md",
             "Particle Swarm Optimization" => "solvers/particle_swarm.md",
             "Quasi-Newton" => "solvers/quasi_Newton.md",
-            "Riemannian augmented Lagrangian method" => "solvers/Riemannian_augmented_Lagrangian_method.md"
             "Stochastic Gradient Descent" => "solvers/stochastic_gradient_descent.md",
             "Subgradient method" => "solvers/subgradient.md",
             "Steihaug-Toint TCG Method" => "solvers/truncated_conjugate_gradient_descent.md",
