+++ src/solvers/augmented_Lagrangian_method.jl
@@ -1,13 +1,81 @@
-# Recreating Changshuo Liu's Matlab source code in Julia
-# original code by Changshuo Liu: https://github.com/losangle/Optimization-on-manifolds-with-extra-constraints/blob/master/solvers/almbddmultiplier.m
-# uses Riemannian limited memory BFGS solver
-# get n_ineq_constraint, n_eq_constraint
+@doc raw"""
+    augmented_Lagrangian_method(M, F, sub_problem, sub_options, n_ineq_constraint, n_eq_constraint)
 
+perform the augmented Lagrangian method (ALM)[^Liu2020]. The aim of the ALM is to find the solution of the `ConstrainedProblem`
+```math
+\begin{aligned}
+\min_{x ∈\mathcal{M}} &f(x)\\
+\text{subject to } &g_i(x)\leq 0 \quad ∀ i= 1, …, m,\\
+\quad &h_j(x)=0 \quad ∀ j=1,…,p,
+\end{aligned}
+```
+where `M` is a Riemannian manifold, and ``f``, ``\{g_i\}_{i=1}^m`` and ``\{h_j\}_{j=1}^p`` are twice continuously differentiable functions from `M` to ℝ.
+For that, in every step ``k`` of the algorithm, the augmented Lagrangian function
+```math
+\mathcal{L}_{ρ^{(k-1)}}(x, λ^{(k-1)}, γ^{(k-1)}) = f(x) + \frac{ρ^{(k-1)}}{2} (\sum_{j=1}^p (h_j(x)+\frac{γ_j^{(k-1)}}{ρ^{(k-1)}})^2 + \sum_{i=1}^m \max\left\{0,\frac{λ_i^{(k-1)}}{ρ^{(k-1)}}+ g_i(x)\right\}^2)
+```
+is minimized over all ``x ∈\mathcal{M}``, where ``λ^{(k-1)}`` and ``γ_j^{(k-1)}`` are the current iterations of the Lagrange multipliers and ``ρ^{(k-1)}`` is the current penalty parameter.
+
+Then, the Lagrange multipliers are updated by 
+```math
+γ_j^{(k)} =\operatorname{clip}_{[γ_j^{\min},γ_j^{\max}]} (γ_j^{(k-1)} + ρ^{(k-1)} h_j(x^{(k)})) \text{for all} j=1,…,p
+```
+and
+```math
+λ_i^{(k)} =\operatorname{clip}_{[0,λ_i^{\max}]} (λ_i^{(k-1)} + ρ^{(k-1)} g_i(x^{(k)}))
+```
+for all ``i=1,…,m``, where ``γ_j^{\min} \leq γ_j^{\max}`` and ``λ_i^{\max}`` are the multiplier boundaries. ###### check how they are implemented
+
+Next, we update the accuracy tolerance ``ϵ`` by setting
+```math
+ϵ^{(k)}=\max\{ϵ_{\min}, θ_ϵ ϵ^{(k-1)}\},
+
+```
+where ``ϵ_{\min}`` is the lowest value ``ϵ`` is allowed to become and ``θ_ϵ ∈ (0,1)`` is constant scaling factor.
+
+[^Liu2020]:
+    > C. Liu, N. Boumal, __Simple Algorithms for Optimization on Riemannian Manifolds with Constraints__,
+    > In: Applied Mathematics & Optimization, vol 82, 949–981 (2020),
+    > doi [10.1007/s00245-019-09564-3](https://doi.org/10.1007/s00245-019-09564-3)
+
+
+# Input
+* `M` – a manifold ``\mathcal M``
+* `F` – a cost function ``F:\mathcal M→ℝ`` to minimize
+* `sub_problem` – problem for the subsolver
[…]
