### 2021-11-16 b6bc76c
Added check for extension in the 'linesearch_backtrack'.
