### 2021-11-22 3988474
adapts the changes talked about in the last meeting
