### 2021-11-26 JuliaManifolds/Manopt.jl#96 @kellertuer
Have you had the time to check whether all points were addressed/solved?
