### 2021-11-27 #3866 @krystophny
@dpsanders @kellertuer I have completed my review of the `Manopt.jl` package and the software paper. Please forgive the long time it took, despite the optimism on finishing sooner. After a detailed analysis, I haven't found any obvious requirements to improve, so I can recommend publication in the present form.

In my opinion, software and paper are of high quality and up to standards in terms of scientific methods, design, documentation and testing. Compared to the existing related `Manopt` packages in Matlab and Python, `Manopt.jl` has been redesigned from ground up, runs on the high-performance Julia language, and implements novel algorithms in the field from the author's active field of research. I think the package looks highly useful, and could also be interesting to apply in my field of research (stellarator optimization).
