### 2021-12-01 4ecc3cd
Update to Julia 1.7, check structure and run Formatter.
