### 2021-12-01 c795392
Switch notebook to use Pluto Package manager, try without the JS trick.
