+++ docs/make.jl
@@ -68,21 +68,6 @@ for (i, f) in enumerate(pluto_files)
         }
         </style>
         $html
-        <script>
-        require(['jquery', 'katex', 'katex-auto-render'], function(\$, katex, renderMathInElement) {
-            \$(document).ready(function() {
-                renderMathInElement(document.body);
-                /* Apparently, Pluto wraps LaTeX inside a span class 'tex'. */
-                var equations = \$('p.tex');
-                for (var i = 0; i < equations.length; i++) {
-                    var text = equations[i].textContent;
-                    text = text.split('\$').join('');
-                    console.log(text);
-                    render(text, equations[i]);
-                }
-            })
-        })
-        </script>
         ```
         """,
     )
