### 2021-12-01 2c54fcb
correct edit url and re-add script for rendering.
