### 2021-12-01 abcbf93
Update docs/make.jl
Co-authored-by: Rik Huijzer <t.h.huijzer@rug.nl>
