### 2021-12-01 c9b9cbf
Updated documentation plan and solver
