+++ src/plans/higher_order_primal_dual_plan.jl
@@ -1,7 +1,7 @@
 # TODO: change documentation
 # DONE: addapted struct and functions
 @doc raw"""
-    PrimalDualProblem {mT <: AbstractManifold, nT <: AbstractManifold} <: PrimalDualProblem} <: Problem
+    PrimalDualSemismoothNewtonProblem {mT <: AbstractManifold, nT <: AbstractManifold} <: PrimalDualProblem} <: Problem
 
 Describes a Problem for the linearized Chambolle-Pock algorithm.
 
@@ -9,17 +9,18 @@ Describes a Problem for the linearized Chambolle-Pock algorithm.
 
 * `M`, `N` – two manifolds $\mathcal M$, $\mathcal N$
 * `cost` $F + G(Λ(⋅))$ to evaluate interims cost function values
-* `forward_oprator` the operator for the forward operation in the algorthm, either $Λ$ (exact) or $DΛ$ (linearized).
+* `linearized_operator` the linearization $DΛ(⋅)[⋅]$ of the operator $Λ(⋅)$.
 * `linearized_adjoint_operator` The adjoint differential $(DΛ)^* \colon \mathcal N \to T\mathcal M$
 * `prox_F` the proximal map belonging to $f$
+* `diff_prox_F` the (Clarke Generalized) differential of the proximal maps of $F$
 * `prox_G_dual` the proximal map belonging to $g_n^*$
-* `Λ` – (`fordward_operator`) for the linearized variant, this has to be set to the exact forward operator.
-  This operator is required in several variants of the linearized algorithm.
-  Since the exact variant is the default, `Λ` is by default set to `forward_operator`.
+* `diff_prox_dual_G` the (Clarke Generalized) differential of the proximal maps of $G^\ast_n$
+* `Λ` – the exact forward operator.
+  This operator is required if `Λ(m)=n` does not hold.
 
 # Constructor
 
-    LinearizedPrimalDualProblem(M, N, cost, prox_F, prox_G_dual, forward_operator, adjoint_linearized_operator,Λ=forward_operator)
+    PrimalDualSemismoothNewtonProblem(M, N, cost, prox_F, prox_G_dual, forward_operator, adjoint_linearized_operator,Λ)
 
 """
 mutable struct PrimalDualSemismoothNewtonProblem{
@@ -69,16 +70,9 @@ end
 * `n` - base point on $ \mathcal N $
 * `x` - an initial point on $x^{(0)} \in \mathcal M$ (and its previous iterate)
 * `ξ` - an initial tangent vector $\xi^{(0)}\in T^*\mathcal N$ (and its previous iterate)
-* `xbar` - the relaxed iterate used in the next dual update step (when using `:primal` relaxation)
-* `ξbar` - the relaxed iterate used in the next primal update step (when using `:dual` relaxation)
-* `Θ` – factor to damp the helping $\tilde x$
 * `primal_stepsize` – (`1/sqrt(8)`) proximal parameter of the primal prox
 * `dual_stepsize` – (`1/sqrt(8)`) proximnal parameter of the dual prox
-* `acceleration` – (`0.`) acceleration factor due to Chambolle & Pock
-* `relaxation` – (`1.`) relaxation in the primal relaxation step (to compute `xbar`)
-* `relax` – (`_primal`) which variable to relax (`:primal` or `:dual`)
 * `stop` - a [`StoppingCriterion`](@ref)
-* `type` – (`exact`) whether to perform an `:exact` or `:linearized` Chambolle-Pock
 * `update_primal_base` (`(p,o,i) -> o.m`) function to update the primal base
 * `update_dual_base` (`(p,o,i) -> o.n`) function to update the dual base
[…]
