### 2021-12-01 7082b47
Update docs/make.jl
Co-authored-by: Rik Huijzer <t.h.huijzer@rug.nl>
