+++ docs/make.jl
@@ -60,6 +60,7 @@ for (i, f) in enumerate(pluto_files)
         ```@meta
         EditURL = "$(pluto_src_folder)$(f).jl"
         ```
+
         ```@raw html
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/fonsp/Pluto.jl@0.16.4/frontend/treeview.css" type="text/css" />
         <style>
