+++ pluto/HigherOrderDev.jl
@@ -1,0 +1,15 @@
+### A Pluto.jl notebook ###
+# v0.17.2
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ 4a1dd24e-52c6-11ec-20a6-3d677b47806f
+begin
+	using Pkg
+	Pkg.activate() # use global enrivonment of current Julia, esp Manopt in dev mode for now
+	using Manifolds, Manopt
+end
+
+# ╔═╡ Cell order:
+# ╠═4a1dd24e-52c6-11ec-20a6-3d677b47806f
