### 2021-12-01 3d582fe
try to switch back to the older JS.
