+++ docs/make.jl
@@ -70,25 +70,20 @@ for (i, f) in enumerate(pluto_files)
         }
         </style>
         $html
-        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/katex@0.15.1/dist/katex.min.css" integrity="sha384-R4558gYOUz8mP9YWpZJjofhk+zx0AS11p36HnD2ZKj/6JR5z27gSSULCNHIRReVs" crossorigin="anonymous">
-        <!-- The loading of KaTeX is deferred to speed up page rendering -->
-        <script defer src="https://cdn.jsdelivr.net/npm/katex@0.15.1/dist/katex.min.js" integrity="sha384-z1fJDqw8ZApjGO3/unPWUPsIymfsJmyrDVWC8Tv/a1HeOtGmkwNd/7xUS0Xcnvsx" crossorigin="anonymous"></script>
-        <!-- To automatically render math in text elements, include the auto-render extension: -->
-        <script defer src="https://cdn.jsdelivr.net/npm/katex@0.15.1/dist/contrib/auto-render.min.js" integrity="sha384-+XBljXPPiv+OzfbB3cVmLHf4hdUFHlWNZN5spNQ7rmHTXpd7WvJum6fIACpNNfIR" crossorigin="anonymous" onload="renderMathInElement(document.body);"></script>
         <script>
-            const options = {
-                delimiters: [
-                    {left: "\$\$", right: "\$\$", display: true},
-                    {left: "\$", right: "\$", display: false},
-                    {left: "\\begin{equation}", right: "\\end{equation}", display: true},
-                    {left: "\\begin{align}", right: "\\end{align}", display: true},
-                    {left: "\\begin{alignat}", right: "\\end{alignat}", display: true},
-                    {left: "\\begin{gather}", right: "\\end{gather}", display: true},
-                    {left: "\\(", right: "\\)", display: false},
-                    {left: "\\[", right: "\\]", display: true}
-                ]
-            };
-            renderMathInElement(document.body, options);
+        require(['jquery', 'katex', 'katex-auto-render'], function(\$, katex, renderMathInElement) {
+            \$(document).ready(function() {
+                renderMathInElement(document.body);
+                /* Apparently, Pluto wraps LaTeX inside a span class 'tex'. */
+                var equations = \$('p.tex');
+                for (var i = 0; i < equations.length; i++) {
+                    var text = equations[i].textContent;
+                    text = text.split('\$').join('');
+                    console.log(text);
+                    render(text, equations[i]);
+                }
+            })
+        })
         </script>
         ```
         """,
