### 2021-12-01 f442107
Raise compat to 0.7
