+++ docs/Project.toml
@@ -18,5 +18,5 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
 Documenter = "0.24"
-Manifolds = "0.5"
+Manifolds = "0.5,0.6,0.7"
 ManifoldsBase = "0.11,0.12"
