### 2021-12-01 23e4be2
first try on using jQuery to style output, but though JS is loaded it says $ unknown and I do not know why.
