+++ pluto/HigherOrderDev.jl
@@ -8,8 +8,395 @@ using InteractiveUtils
 begin
 	using Pkg
 	Pkg.activate() # use global enrivonment of current Julia, esp Manopt in dev mode for now
-	using Manifolds, Manopt
+	using Manifolds, Manopt, LinearAlgebra
 end
 
+
+# ╔═╡ 607c1fbf-0b58-4dc0-a0eb-1a682f97b4f7
+md"""
+# Using Primal-Dual Riemannian Semismooth Newton for ``S^2``-valued data
+"""
+
+# ╔═╡ 4522df74-6c30-4dde-bb3e-10d46be06c6c
+begin
+	experiment_name = "S2_Signal_TV_PDRSSN"
+	export_orig = true
+	export_primal = false
+	export_table = false
+	use_debug = true
+	#
+	# Automatic Script Settings
+	current_folder = @__DIR__
+	export_any = export_orig || export_primal || export_table
+	results_folder = joinpath(current_folder, "Signal_TV")
+	# Create folder if we have an export
+	(export_any && !isdir(results_folder)) && mkdir(results_folder)
+
+end
+
+# ╔═╡ 7d3d39db-6791-4565-94c8-18a1959e16c6
+md"""
+# Settings
+"""
+
+# ╔═╡ 830dff11-07d0-4ba3-9de1-f35413c98cc2
+signal_section_size = 4
+
+# ╔═╡ 44b78430-91d4-4157-8431-f6cb0ad7aae8
+α = 5.0
+
+# ╔═╡ 68f01f9d-e5c4-45f4-aa75-1560d1948b54
+σ = 0.5
+
+# ╔═╡ ee68439a-1ce8-4f74-b721-500859eca196
+τ = 0.5
+
+# ╔═╡ 7a0460f1-afe5-46dc-a425-53616f4fb250
+max_iterations = 50
+
+# ╔═╡ 51f3b7b2-c9a7-42c6-a4bc-88a446aeb967
+pixelM = Sphere(2)
+
+# ╔═╡ 621077a4-2b6e-4e3c-8b51-ee3613d9c639
+md"""
+# Generate a signal
+"""
+
+# ╔═╡ a0738ce7-46d1-46ad-8c48-85d600a1c5de
+base = [1.0, 0.0, 0.0]
+
+# ╔═╡ 555fc740-9434-42c8-a647-66512adc8cff
+X = π / 4 * [0.0, 1.0, 0.0]
+
+# ╔═╡ 9b5f4ed8-b91f-45a7-ade0-d3828d7c0e0f
+p1 = exp(pixelM, base, X)
+
+# ╔═╡ 894a1d46-daae-4f15-b26e-dbf61a8a3e97
+p2 = exp(pixelM, base, -X)
+
+# ╔═╡ c4dbabb1-2b13-49de-81d6-be914a3e4ac5
+f = vcat(fill(p1, signal_section_size), fill(p2, signal_section_size))
+
+# ╔═╡ 95d6a845-9050-4bfe-b004-a8603c8271b8
+md"""
+# Model TV for a given pixelManifold
+"""
+
+# ╔═╡ 3ec756a4-491e-41f6-9e66-2616971a81ad
+rep(d) = (d > 1) ? [ones(Int, d)..., d] : d
+
+# ╔═╡ b703566e-aba2-4d7d-8982-d258ec44c837
+res(s::Int) = res([s])
+
+# ╔═╡ b69bf0b3-94c5-447b-9ff6-fdb4e37876fc
+res(s) = (length(s) > 1) ? [s..., length(s)] : s
+
+# ╔═╡ 94f094e0-b59c-4f5f-8361-f4ce259b3cba
[…]
