### 2021-12-02 cfca20c
switch back to local env after Pluto runs.
