### 2021-12-02 e1b6987
Move benchmark to a pluto notebook.
