+++ pluto/Benchmark.jl
@@ -1,0 +1,653 @@
+### A Pluto.jl notebook ###
+# v0.17.2
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ 62b0cf82-e00c-498b-a6d8-124ea9950ef4
+using Manopt, Manifolds, Random, BenchmarkTools
+
+# ╔═╡ d5e636f0-533a-11ec-168a-e118d53012cb
+md"""
+# Illustration how to use mutating gradient functions
+
+When it comes to time critital operations, a main ingredient in Julia are
+mutating functions, i.e. those that compute in place without additional Memory
+allocations. In the following the illustrate how to do this with `Manopt.jl`.
+
+Let's start with the same function as in [Get Started: Optimize!](https://manoptjl.org/stable/tutorials/MeanAndMedian.html)
+and compute the mean of some points. Just that here we use the sphere ``\mathbb S^{30}``
+and `n=800` points.
+
+From the just mentioned example, the implementation looks like
+"""
+
+# ╔═╡ 967e5caa-1d27-4c24-a882-6f7126053754
+begin
+	Random.seed!(42)
+	m = 30
+	M = Sphere(m)
+	n = 800
+	σ = π / 8
+	x = zeros(Float64, m + 1)
+	x[2] = 1.0
+	data = [exp(M, x, random_tangent(M, x, Val(:Gaussian), σ)) for i in 1:n]
+end;
+
+# ╔═╡ 1436cdc7-1e84-4438-937d-2211856348de
+md"""
+## Classical definition
+
+The variant from the previous tutorial defines a cost ``F(x)`` and its gradient ``gradF(x)``
+"""
+
+# ╔═╡ 6f1b04cf-2a76-4660-916c-3b8b8472d3c0
+F(x) = sum(1 / (2 * n) * distance.(Ref(M), Ref(x), data) .^ 2)
+
+# ╔═╡ e487f38c-334e-4819-b775-a9a44632d7ff
+gradF(M, x) = sum(1 / n * grad_distance.(Ref(M), data, Ref(x)))
+
+# ╔═╡ e4a7b318-65e7-40bb-899d-8bdeeefbf304
+md"""
+we further set the stopping criterion to be a little more strict, then we obtain
+"""
+
+# ╔═╡ 5a2d6d9f-20dc-41a2-bfce-32509a1ef106
+begin
+	sc = StopWhenGradientNormLess(1e-10)
+	x0 = random_point(M)
+	m1 = gradient_descent(M, F, gradF, x0; stopping_criterion=sc)
+
+	@benchmark gradient_descent($M, $F, $gradF, $x0; stopping_criterion=$sc)
+end
+
+# ╔═╡ 306c7706-f399-4104-9159-c2e6b0bca189
+md"""
+## Inplace computation of the gradient
+
+We can reduce the memory allocations, by implementing the gradient as a [functor](https://docs.julialang.org/en/v1/manual/methods/#Function-like-objects).
+The motivation is twofold: On the one hand, we want to avoid variables from global scope,
+for example the manifold `M` or the `data`, to be used within the function
[…]
