### 2021-12-02 1bed477
remove randomness from a quasi newton test.
