+++ test/solvers/test_quasi_Newton.jl
@@ -193,8 +193,7 @@ Random.seed!(42)
         n = 1000
         k = 5
         M = Stiefel(n, k)
-        A = randn(n, n)
-        A = (A + A') / 2
+        A = one(zeros(n, n))
         F(::Stiefel, X) = tr((X' * A * X) * Diagonal(k:-1:1))
         gradF = GradF(A, Diagonal(Float64.(collect(k:-1:1))))
 
@@ -208,7 +207,7 @@ Random.seed!(42)
             vector_transport_method=ProjectionTransport(),
             retraction_method=QRRetraction(),
             cautious_update=true,
-            stopping_criterion=StopWhenGradientNormLess(10^(-6)),
+            stopping_criterion=StopWhenGradientNormLess(10^(-4)),
         )
 
         x_inverseBFGSHuang = quasi_Newton(
@@ -221,8 +220,8 @@ Random.seed!(42)
             vector_transport_method=ProjectionTransport(),
             retraction_method=QRRetraction(),
             cautious_update=true,
-            stopping_criterion=StopWhenGradientNormLess(10^(-6)),
+            stopping_criterion=StopWhenGradientNormLess(10^(-4)),
         )
-        @test isapprox(M, x_inverseBFGSCautious, x_inverseBFGSHuang; atol=2e-5)
+        @test isapprox(M, x_inverseBFGSCautious, x_inverseBFGSHuang; atol=2e-4)
     end
 end
