+++ test/functions/test_adjoint_differentials.jl
@@ -35,15 +35,11 @@ using Manifolds, Manopt, Test, ManifoldsBase
     qP = [p, p, q]
     XP = [X, zero_vector(M, p), -X]
     YP = similar.(XP)
-    @test norm(
-        Mp,
-        pP,
-        adjoint_differential_forward_logs(Mp, pP, XP) -
-        [-X, zero_vector(M, p), zero_vector(M, p)],
-    ) ≈ 0 atol = 4 * 10.0^(-16)
-    adjoint_differential_forward_logs!(Mp, YP, pP, XP)
-    @test norm(Mp, pP, YP - [-X, zero_vector(M, p), zero_vector(M, p)]) ≈ 0 atol =
+    ZP = adjoint_differential_forward_logs(Mp, pP, XP)
+    @test norm(Mp, pP, ZP - [-X, zero_vector(M, p), zero_vector(M, p)]) ≈ 0 atol =
         4 * 10.0^(-16)
+    adjoint_differential_forward_logs!(Mp, YP, pP, XP)
+    @test isapprox(Mp, pP, YP, ZP; atol=4 * 10.0^(-16))
     ZP = [[0.0, π / 2, 0.0], [0.0, 0.0, 0.0], [π / 2, 0.0, 0.0]]
     @test adjoint_differential_log_argument(Mp, pP, qP, XP) == ZP
     adjoint_differential_log_argument!(Mp, YP, pP, qP, XP)
