+++ .github/workflows/ci.yml
@@ -27,4 +27,4 @@ jobs:
           file: ./lcov.info
           name: codecov-umbrella
           fail_ci_if_error: false
-        if: ${{ matrix.julia-version == '1.6' && matrix.os =='ubuntu-latest' }}
+        if: ${{ matrix.julia-version == '1.7' && matrix.os =='ubuntu-latest' }}
