+++ src/solvers/augmented_Lagrangian_method.jl
@@ -1,5 +1,5 @@
 @doc raw"""
-    augmented_Lagrangian_method(M, F, sub_problem, sub_options, n_ineq_constraint, n_eq_constraint)
+    augmented_Lagrangian_method(M, F, gradF, sub_problem, sub_options, G, H, gradG, gradH)
 
 perform the augmented Lagrangian method (ALM)[^Liu2020]. The aim of the ALM is to find the solution of the `ConstrainedProblem`
 ```math
@@ -22,16 +22,28 @@ Then, the Lagrange multipliers are updated by
 ```
 and
 ```math
-λ_i^{(k)} =\operatorname{clip}_{[0,λ_i^{\max}]} (λ_i^{(k-1)} + ρ^{(k-1)} g_i(x^{(k)}))
+λ_i^{(k)} =\operatorname{clip}_{[0,λ_i^{\max}]} (λ_i^{(k-1)} + ρ^{(k-1)} g_i(x^{(k)})) \text{for all}  i=1,…,m,
 ```
-for all ``i=1,…,m``, where ``γ_j^{\min} \leq γ_j^{\max}`` and ``λ_i^{\max}`` are the multiplier boundaries. ###### check how they are implemented
+where ``γ_j^{\min} \leq γ_j^{\max}`` and ``λ_i^{\max}`` are the multiplier boundaries. 
 
 Next, we update the accuracy tolerance ``ϵ`` by setting
 ```math
 ϵ^{(k)}=\max\{ϵ_{\min}, θ_ϵ ϵ^{(k-1)}\},
-
 ```
 where ``ϵ_{\min}`` is the lowest value ``ϵ`` is allowed to become and ``θ_ϵ ∈ (0,1)`` is constant scaling factor.
+
+Last, we update the penalty parameter ``ρ``. For this, we define
+```math
+σ^{(k)}=\max_{j=1,…,p, i=1,…,m} \{\|h_j(x^{(k)})\|, \|\max_{i=1,…,m}\{g_i(x^{(k)}), -\frac{λ_i^{(k-1)}}{ρ^{(k-1)}} \}\| \}.
+```
+Then, we update `ρ` according to
+```math
+ρ^{(k)} = \begin{cases}
+ρ^{(k-1)}/θ_ρ,  & \text{if } σ^{(k)}\leq θ_ρ σ^{(k-1)} ,\\
+ρ^{(k-1)}, & \text{else,}
+\end{cases}
+```
+where ``θ_ρ \in (0,1)`` is a constant scaling factor.
 
 [^Liu2020]:
     > C. Liu, N. Boumal, __Simple Algorithms for Optimization on Riemannian Manifolds with Constraints__,
@@ -42,30 +54,33 @@ where ``ϵ_{\min}`` is the lowest value ``ϵ`` is allowed to become and ``θ_ϵ 
 # Input
 * `M` – a manifold ``\mathcal M``
 * `F` – a cost function ``F:\mathcal M→ℝ`` to minimize
+* `gradF` – the gradient of the cost function
 * `sub_problem` – problem for the subsolver
 * `sub_options` – options of the subproblem
-* `n_ineq_constraint` – the number of inequality constraints of the problem
-* `n_eq_constraint` – the number of equality constraints of the problem
+* `G` – the inequality constraints
+* `H` – the equality constraints 
+* `gradG` – the gradient of the inequality constraints
+* `gradH` – the gradient of the equality constraints
 
 # Optional
 * `x` - initial point
-* `max_inner_iter` - (`200`) the maximum number of iterations the subsolver should perform in each iteration ##### 
+* `max_inner_iter` - (`200`) the maximum number of iterations the subsolver should perform in each iteration 
 * `num_outer_itertgn` - (`30`)
[…]
