### 2021-12-02 6bb9bf5
attempt solving bug types in higher order plan
