### 2021-12-02 bf2ce16
start the Bezier pluto notebook.
