+++ pluto/bezier.jl
@@ -1,0 +1,1158 @@
+### A Pluto.jl notebook ###
+# v0.17.2
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ 6f9be298-5374-11ec-3b9d-fd74a32d6b53
+using Colors, Images, Manopt, Manifolds
+
+# ╔═╡ e701b513-4ad9-474c-9d79-60a24f545ebd
+md"""
+# Bezier curves and their acceleration
+
+This tutorial illustrates how Bézier curves are generalized to manifolds and how to
+minimize their acceleration, i.e. how to get a curve that is as straight or as geodesic
+while fulfilling constraints
+
+This example also illustrates how to apply the minimization on the corresponding [`PowerManifold`](https://juliamanifolds.github.io/Manifolds.jl/latest/manifolds/power.html) manifold using a [`gradient_descent`](https://manoptjl.org/stable/solvers/gradient_descent.html) with [`ArmijoLinesearch`](https://manoptjl.org/stable/plans/index.html#Manopt.ArmijoLinesearch).
+"""
+
+# ╔═╡ 087fb979-d68d-4ee7-9be5-fad992fdb792
+md"and we define some colors from [Paul Tol](https://personal.sron.nl/~pault/)"
+
+# ╔═╡ 85d8985b-f5ee-4cba-bcc5-528049798330
+begin
+	black = RGBA{Float64}(colorant"#000000")
+	TolVibrantBlue = RGBA{Float64}(colorant"#0077BB")
+	TolVibrantOrange = RGBA{Float64}(colorant"#EE7733")
+	TolVibrantMagenta = RGBA{Float64}(colorant"#EE3377")
+	TolVibrantCyan = RGBA{Float64}(colorant"#33BBEE")
+	TolVibrantTeal = RGBA{Float64}(colorant"#009988")
+end;
+
+# ╔═╡ e62e36e9-19ce-4a7d-8bda-5634813c1ee9
+md"Let's also set up a few parameters"
+
+# ╔═╡ 18d378dd-abc6-4fa6-a2c9-1cc09e4cec46
+begin
+	geo_pts = collect(range(0.0, 1.0; length=101))
+	bezier_pts = collect(range(0.0, 3.0; length=201))
+	camera_position = (-1.0, -0.7, 0.3)
+	#render asy yes/no. If not, images included w/ markdown are assumed to be prerendered
+	render_asy = true
+	image_prefix = render_asy ? "bezier/" : "../assets/images/"
+end;
+
+# ╔═╡ dee4eb65-145c-4b2b-98a7-ae3a7045f24b
+md"""
+We finally load our data, see [`artificial_S2_composite_bezier_curve`](https://manoptjl.org/stable/helpers/data.html#Manopt.artificial_S2_composite_bezier_curve-Tuple{}), a composite Bezier curve consisting of 3 segments on the Sphere. The middle segment consists of the control points
+"""
+
+# ╔═╡ 2b485956-459d-40e7-9364-79ea77073804
+B = artificial_S2_composite_bezier_curve(); b = B[2].pts
+
+# ╔═╡ 98c925ac-9ac8-42f3-9855-a01e5eaac350
+M = Sphere(2)
+
+# ╔═╡ d22e825e-d8ff-44fe-8b06-b14c09765180
+md"""
+On Euclidean spaces Bézier curves of these ``n=4`` so called control points like this segment yield polynomials of degree``3``.
[…]
