### 2021-12-02 fd0d67d
Yes we need exactly that, _and_ \lambda is a kw-parameter so we have to pass it that way (and lets pass eval as well).
