+++ src/Manopt.jl
@@ -199,8 +199,12 @@ export get_cost,
     get_gradients!,
     get_primal_prox,
     get_primal_prox!,
+    get_differential_primal_prox,
+    get_differential_primal_prox!,
     get_dual_prox,
     get_dual_prox!,
+    get_differential_dual_prox,
+    get_differential_dual_prox!,
     linearized_forward_operator,
     linearized_forward_operator!,
     adjoint_linearized_operator,
