+++ src/plans/primal_dual_plan.jl
@@ -80,8 +80,8 @@ function PrimalDualProblem(
 end
 
 @doc raw"""
-    y = get_primal_prox(p::PrimalDualProblem, σ, x)
-    get_primal_prox!(p::PrimalDualProblem, y, σ, x)
+    y = get_primal_prox(p::AbstractPrimalDualProblem, σ, x)
+    get_primal_prox!(p::AbstractPrimalDualProblem, y, σ, x)
 
 Evaluate the proximal map of ``F`` stored within [`PrimalDualProblem`](@ref)
 
@@ -93,23 +93,23 @@ which can also be computed in place of `y`.
 """
 get_primal_prox(::AbstractPrimalDualProblem, ::Any...)
 
-function get_primal_prox(p::AbstractPrimalDualProblem{AllocatingEvaluation}, σ, x)
+function get_primal_prox(p::AbstractPrimalDualProblem{<:AllocatingEvaluation}, σ, x)
     return p.prox_F!!(p.M, σ, x)
 end
-function get_primal_prox(p::AbstractPrimalDualProblem{MutatingEvaluation}, σ, x)
+function get_primal_prox(p::AbstractPrimalDualProblem{<:MutatingEvaluation}, σ, x)
     y = allocate_result(p.M, get_primal_prox, x)
     return p.prox_F!!(p.M, y, σ, x)
 end
-function get_primal_prox!(p::AbstractPrimalDualProblem{AllocatingEvaluation}, y, σ, x)
+function get_primal_prox!(p::AbstractPrimalDualProblem{<:AllocatingEvaluation}, y, σ, x)
     return copyto!(p.M, y, p.prox_F!!(p.M, σ, x))
 end
-function get_primal_prox!(p::AbstractPrimalDualProblem{MutatingEvaluation}, y, σ, x)
+function get_primal_prox!(p::AbstractPrimalDualProblem{<:MutatingEvaluation}, y, σ, x)
     return p.prox_F!!(p.M, y, σ, x)
 end
 
 @doc raw"""
-    y = get_dual_prox(p::PrimalDualProblem, n, τ, ξ)
-    get_dual_prox!(p::PrimalDualProblem, y, n, τ, ξ)
+    y = get_dual_prox(p::AbstractPrimalDualProblem, n, τ, ξ)
+    get_dual_prox!(p::AbstractPrimalDualProblem, y, n, τ, ξ)
 
 Evaluate the proximal map of ``G_n^*`` stored within [`PrimalDualProblem`](@ref)
 
@@ -121,17 +121,17 @@ which can also be computed in place of `y`.
 """
 get_dual_prox(::AbstractPrimalDualProblem, ::Any...)
 
-function get_dual_prox(p::AbstractPrimalDualProblem{AllocatingEvaluation}, n, τ, ξ)
+function get_dual_prox(p::AbstractPrimalDualProblem{<:AllocatingEvaluation}, n, τ, ξ)
     return p.prox_G_dual!!(p.N, n, τ, ξ)
 end
-function get_dual_prox(p::AbstractPrimalDualProblem{MutatingEvaluation}, n, τ, ξ)
+function get_dual_prox(p::AbstractPrimalDualProblem{<:MutatingEvaluation}, n, τ, ξ)
     η = allocate_result(p.N, get_dual_prox, ξ)
     return p.prox_G_dual!!(p.N, η, n, τ, ξ)
 end
-function get_dual_prox!(p::AbstractPrimalDualProblem{AllocatingEvaluation}, η, n, τ, ξ)
+function get_dual_prox!(p::AbstractPrimalDualProblem{<:AllocatingEvaluation}, η, n, τ, ξ)
     return copyto!(p.N, η, p.prox_G_dual!!(p.N, n, τ, ξ))
 end
[…]
