### 2021-12-02 724eb15
Nearly (up one bug) finishes Bezier Pluto Notebook.
