### 2021-12-02 48ab0c7
Finish the Bezier notebook.
