### 2021-12-02 b1b2cb8
Trying to use images in PlutoStaticHTML.
