+++ pluto/bezier.jl
@@ -23,12 +23,12 @@ md"and we define some colors from [Paul Tol](https://personal.sron.nl/~pault/)"
 
 # ╔═╡ 85d8985b-f5ee-4cba-bcc5-528049798330
 begin
-	black = RGBA{Float64}(colorant"#000000")
-	TolVibrantBlue = RGBA{Float64}(colorant"#0077BB")
-	TolVibrantOrange = RGBA{Float64}(colorant"#EE7733")
-	TolVibrantMagenta = RGBA{Float64}(colorant"#EE3377")
-	TolVibrantCyan = RGBA{Float64}(colorant"#33BBEE")
-	TolVibrantTeal = RGBA{Float64}(colorant"#009988")
+    black = RGBA{Float64}(colorant"#000000")
+    TolVibrantBlue = RGBA{Float64}(colorant"#0077BB")
+    TolVibrantOrange = RGBA{Float64}(colorant"#EE7733")
+    TolVibrantMagenta = RGBA{Float64}(colorant"#EE3377")
+    TolVibrantCyan = RGBA{Float64}(colorant"#33BBEE")
+    TolVibrantTeal = RGBA{Float64}(colorant"#009988")
 end;
 
 # ╔═╡ e62e36e9-19ce-4a7d-8bda-5634813c1ee9
@@ -36,12 +36,12 @@ md"Let's also set up a few parameters"
 
 # ╔═╡ 18d378dd-abc6-4fa6-a2c9-1cc09e4cec46
 begin
-	geo_pts = collect(range(0.0, 1.0; length=101))
-	bezier_pts = collect(range(0.0, 3.0; length=201))
-	camera_position = (-1.0, -0.7, 0.3)
-	#render asy yes/no. If not, images included w/ markdown are assumed to be prerendered
-	render_asy = true
-	image_prefix = render_asy ? "bezier/" : "../assets/images/"
+    geo_pts = collect(range(0.0, 1.0; length=101))
+    bezier_pts = collect(range(0.0, 3.0; length=201))
+    camera_position = (-1.0, -0.7, 0.3)
+    #render asy yes/no. If not, images included w/ markdown are assumed to be prerendered
+    render_asy = false
+    image_prefix = String(@__DIR__)*"/bezier"
 end;
 
 # ╔═╡ dee4eb65-145c-4b2b-98a7-ae3a7045f24b
@@ -50,7 +50,8 @@ We finally load our data, see [`artificial_S2_composite_bezier_curve`](https://m
 """
 
 # ╔═╡ 2b485956-459d-40e7-9364-79ea77073804
-B = artificial_S2_composite_bezier_curve(); b = B[2].pts
+B = artificial_S2_composite_bezier_curve();
+b = B[2].pts;
 
 # ╔═╡ 98c925ac-9ac8-42f3-9855-a01e5eaac350
 M = Sphere(2)
@@ -66,7 +67,8 @@ Let's evaliuate this at the point ``t=\frac{1}{4}∈[0,1]``. We first compute
 """
 
 # ╔═╡ 157d13c8-7311-4832-a9ff-da22125b2bbf
-t = 0.66; pts1 = shortest_geodesic.(Ref(M), b[1:3], b[2:4], Ref(t))
+t = 0.66;
+pts1 = shortest_geodesic.(Ref(M), b[1:3], b[2:4], Ref(t));
 
 # ╔═╡ 5af58241-50c7-464e-bfa3-53819e605793
 md"""
@@ -75,8 +77,8 @@ We obtain 3 points on the geodesics connecting the control points. Repeating thi
 
[…]
