+++ src/tutorials/BezierCurves.jl
@@ -1,386 +1,0 @@
-# # [Bezier curves and their acceleration](@id BezierCurvesTutorial)
-#
-# This tutorial illustrates how Bézier curves are generalized to manifolds and how to
-# minimize their acceleration, i.e. how to get a curve that is as straight or as geodesic
-# while fulfilling constraints
-#
-# This example also illustrates how to apply the minimization on the corresponding [`PowerManifold`](https://juliamanifolds.github.io/Manifolds.jl/latest/manifolds/power.html) manifold using a [`gradient_descent`](@ref) with [`ArmijoLinesearch`](@ref).
-#
-# ## Table of contents
-# * [Setup](@ref SetupTB)
-# * [de Casteljau algorithm on manifolds](@ref Casteljau)
-# * [Composite Bézire curves](@ref CompositeBezier)
-# * [Minimizing the acceleration of a Bézier curve](@ref MinAccBezier)
-# * [Literature](@ref LiteratureBT)
-#
-# ## [Setup](@id SetupTB)
-#
-# We first initialize the necessary packages
-export_folder = joinpath( #src
-    @__DIR__, #src
-    "..", #src
-    "..", #src
-    "docs", #src
-    "src", #src
-    "assets", #src
-    "images", #src
-    "tutorials", #src
-) #src
-using Manopt, Manifolds
-# and we define some colors from [Paul Tol](https://personal.sron.nl/~pault/)
-using Colors
-black = RGBA{Float64}(colorant"#000000")
-TolVibrantBlue = RGBA{Float64}(colorant"#0077BB")
-TolVibrantOrange = RGBA{Float64}(colorant"#EE7733")
-TolVibrantMagenta = RGBA{Float64}(colorant"#EE3377")
-TolVibrantCyan = RGBA{Float64}(colorant"#33BBEE")
-TolVibrantTeal = RGBA{Float64}(colorant"#009988")
-geo_pts = collect(range(0.0, 1.0; length=101)) #hide
-bezier_pts = collect(range(0.0, 3.0; length=201)) #hide
-camera_position = (-1.0, -0.7, 0.3) #hide
-nothing #hide
-# Then we load our data, see [`artificial_S2_composite_bezier_curve`](@ref), a composite
-# Bezier curve consisting of 3 segments on the Sphere
-B = artificial_S2_composite_bezier_curve();
-#
-# ## [De Casteljau algorithm on manifolds](@id Casteljau)
-# This curve can be evaluated using de [Casteljau's algorithm](https://en.wikipedia.org/wiki/De_Casteljau%27s_algorithm)[^Casteljau1959][^Casteljau1963] named after [Paul de Casteljau](https://en.wikipedia.org/wiki/Paul_de_Casteljau)(*1930).
-# To simplify the idea and understand this algorithm, we will first only look at the points of the first segment
-M = Sphere(2)
-b = B[2].pts
-# On Euclidean spaces Bézier curves of these $n=4$ so called control points like this segment yield polynomials of degree $3$.
-# The resulting curve $\gamma: [0,1] → ℝ^m$ is called [Bezier curve](https://en.wikipedia.org/wiki/Bézier_curve) or Bézier spline and is named after [Piérre Bezier](https://en.wikipedia.org/wiki/Pierre_Bézier) (1910–1999).
-# They can be evaluated by the de Casteljau algorithm by evaluating line segments between points.
[…]
