+++ pluto/bezier.jl
@@ -41,7 +41,7 @@ begin
     camera_position = (-1.0, -0.7, 0.3)
     #render asy yes/no. If not, images included w/ markdown are assumed to be prerendered
     render_asy = false
-    image_prefix = String(@__DIR__)*"/bezier"
+    image_prefix = String(@__DIR__) * "/bezier"
 end;
 
 # ╔═╡ dee4eb65-145c-4b2b-98a7-ae3a7045f24b
