### 2021-12-02 c4458ed
Formatter killed another cell.
