### 2021-12-03 ae5a59c
Let's investigate what the thingy gets set to.
