### 2021-12-03 3a0b082
Trying to trick with extracting the bath from the file and use that relative.
