+++ pluto/Bezier.jl
@@ -41,7 +41,8 @@ begin
     camera_position = (-1.0, -0.7, 0.3)
     #render asy yes/no. If not, images included w/ markdown are assumed to be prerendered
     render_asy = false
-    image_prefix = joinpath(pkgdir(Manopt), "pluto", "bezier")
+    localpath = join(splitpath(@__FILE__)[1:(end - 1)], "/") # remove file to get files folder
+    image_prefix = localpath * "/bezier"
     @info image_prefix
 end;
 
