### 2021-12-12 #3866 @sweichwald
@dpsanders @kellertuer I have now also completed my review.

Some minor comments on the software paper have been addressed (cf. https://github.com/JuliaManifolds/Manopt.jl/issues/96),
software design, development, documentation, and continuous integration follow best practices,
and, in my opinion, `Manopt.jl` is a useful quality piece of research software that I recommend for publication.
