+++ pluto/Bezier.jl
@@ -1,11 +1,11 @@
 ### A Pluto.jl notebook ###
-# v0.17.2
+# v0.17.3
 
 using Markdown
 using InteractiveUtils
 
 # ╔═╡ 6f9be298-5374-11ec-3b9d-fd74a32d6b53
-using Colors, Images, Manopt, Manifolds
+using Colors, PlutoUI, Manopt, Manifolds
 
 # ╔═╡ e701b513-4ad9-474c-9d79-60a24f545ebd
 md"""
@@ -127,7 +127,7 @@ end
 md"""This image summarizes Casteljaus algorithm:"""
 
 # ╔═╡ 38b8f8af-ba81-4d3b-83b7-56b0d41ea708
-load(image_prefix * "/Casteljau-illustr.png")
+PlutoUI.LocalResource(image_prefix * "/Casteljau-illustr.png")
 
 # ╔═╡ d6759a8e-a705-4d95-a8da-e172ada644d8
 md"""
@@ -191,7 +191,7 @@ render_asy && begin
 end
 
 # ╔═╡ 41ba8ac2-5a81-46f5-9269-404cc6a1d287
-load(image_prefix * "/Bezier-composite-curve.png")
+PlutoUI.LocalResource(image_prefix * "/Bezier-composite-curve.png")
 
 # ╔═╡ e7fc05b3-a198-4a64-b62a-43cdf8f56238
 md"""
@@ -246,7 +246,7 @@ render_asy && begin
 end
 
 # ╔═╡ 5157ac76-d359-4f18-883d-8afbe78d3b4d
-load(image_prefix * "/Bezier-composite-curve-gradient.png")
+PlutoUI.LocalResource(image_prefix * "/Bezier-composite-curve-gradient.png")
 
 # ╔═╡ 7640628b-e232-4ba3-94aa-d8a00218ff6a
 md"""
@@ -319,7 +319,7 @@ render_asy && begin
 end
 
 # ╔═╡ 4f35b3e8-90ca-4d21-a377-007f93eb0efd
-load(image_prefix * "/Bezier-IP-Min.png")
+PlutoUI.LocalResource(image_prefix * "/Bezier-IP-Min.png")
 
 # ╔═╡ 203a0538-d608-4d3e-b57d-edec20756300
 md"""
@@ -396,7 +396,7 @@ render_asy && begin
 end
 
 # ╔═╡ db8110e0-d271-424c-b920-817dc1aab0c9
-load(image_prefix * "/Bezier-Appr-Min.png")
+PlutoUI.LocalResource(image_prefix * "/Bezier-Appr-Min.png")
 
 # ╔═╡ 0855f66e-f9af-4c63-b041-424c9f3fd655
 md"""Additionally to the last image, the data points ``d_i`` (junction points of the original curve) are shown in orange, the distance between these and the blue junction points is part of the cost function here."""
@@ -408,6 +408,7 @@ minimizer, i.e. the resulting curve, is closer to the original Bézier junction 
 For small ``λ`` the resting curve is closer to a geodesic and the control points are closer to the curve.
 For ``λ=0`` _any_ (not necessarily shortest) geodesic is a solution and the problem is ill-posed.
 To illustrate the effect of ``λ``, the following image contains 1000 runs for ``λ=10`` in dark currant to ``λ=0.01`` in bright yellow.
+
[…]
