+++ src/plans/plan.jl
@@ -6,10 +6,10 @@ include("options.jl")
 include("debug_options.jl")
 include("record_options.jl")
 
-include("alm_plan.jl")
-include("constrained_plan.jl")
 include("cost_plan.jl")
 include("gradient_plan.jl")
+include("constrained_plan.jl")
+include("alm_plan.jl")
 include("conjugate_gradient_plan.jl")
 include("quasi_newton_plan.jl")
 include("hessian_plan.jl")
