+++ src/solvers/augmented_Lagrangian_method.jl
@@ -1,7 +1,7 @@
 @doc raw"""
     augmented_Lagrangian_method(M, F, gradF, sub_problem, sub_options, G, H, gradG, gradH)
 
-perform the augmented Lagrangian method (ALM)[^Liu2020]. The aim of the ALM is to find the solution of the `ConstrainedProblem`
+perform the augmented Lagrangian method (ALM)[^LiuBoumal2020][^source_code]. The aim of the ALM is to find the solution of the [`ConstrainedProblem`](@ref)
 ```math
 \begin{aligned}
 \min_{x ∈\mathcal{M}} &f(x)\\
@@ -14,7 +14,7 @@ For that, in every step ``k`` of the algorithm, the augmented Lagrangian functio
 ```math
 \mathcal{L}_{ρ^{(k-1)}}(x, λ^{(k-1)}, γ^{(k-1)}) = f(x) + \frac{ρ^{(k-1)}}{2} (\sum_{j=1}^p (h_j(x)+\frac{γ_j^{(k-1)}}{ρ^{(k-1)}})^2 + \sum_{i=1}^m \max\left\{0,\frac{λ_i^{(k-1)}}{ρ^{(k-1)}}+ g_i(x)\right\}^2)
 ```
-is minimized over all ``x ∈\mathcal{M}``, where ``λ^{(k-1)}`` and ``γ_j^{(k-1)}`` are the current iterations of the Lagrange multipliers and ``ρ^{(k-1)}`` is the current penalty parameter.
+is minimized over all ``x ∈\mathcal{M}``, where ``λ^{(k-1)}=[λ_1^{(k-1)}, …, λ_m^{(k-1)}]^T`` and ``γ^{(k-1)}=[γ_1^{(k-1)}, …, γ_p^{(k-1)}]^T`` are the current iterations of the Lagrange multipliers and ``ρ^{(k-1)}`` is the current penalty parameter.
 
 Then, the Lagrange multipliers are updated by 
 ```math
@@ -45,10 +45,15 @@ Then, we update `ρ` according to
 ```
 where ``θ_ρ \in (0,1)`` is a constant scaling factor.
 
-[^Liu2020]:
+[^LiuBoumal2020]:
     > C. Liu, N. Boumal, __Simple Algorithms for Optimization on Riemannian Manifolds with Constraints__,
     > In: Applied Mathematics & Optimization, vol 82, 949–981 (2020),
     > doi [10.1007/s00245-019-09564-3](https://doi.org/10.1007/s00245-019-09564-3)
+
+[^source_code]:
+    > original source code to the paper:
+    > C. Liu, N. Boumal, __Simple Algorithms for Optimization on Riemannian Manifolds with Constraints__,
+    > src: [https://github.com/losangle/Optimization-on-manifolds-with-extra-constraints](https://github.com/losangle/Optimization-on-manifolds-with-extra-constraints)
 
 
 # Input
@@ -68,9 +73,12 @@ where ``θ_ρ \in (0,1)`` is a constant scaling factor.
 * `num_outer_itertgn` - (`30`)
 * `ϵ` - (`1e-3`) the accuracy tolerance
 * `ϵ_min` - (`1e-6`) the lower bound for the accuracy tolerance
-* `bound` - (`20`) a bound for the Lagrange multiplier
-* `λ` - (`ones(n_ineq_constraint)`) the Lagrange multiplier with respect to the inequality constraints
-* `γ` - (`ones(n_eq_constraint)`) the Lagrange multiplier with respect to the equality constraints
+#### * `bound` - (`20`) a bound for the Lagrange multiplier
+* `γ_max` - (`20`)
+* `γ_min` - (`- γ_max`)
+* `λ_max` - (`20`)
[…]
