+++ src/solvers/augmented_Lagrangian_method.jl
@@ -18,13 +18,13 @@ is minimized over all ``x ∈\mathcal{M}``, where ``λ^{(k-1)}=[λ_1^{(k-1)}, �
 
 Then, the Lagrange multipliers are updated by 
 ```math
-γ_j^{(k)} =\operatorname{clip}_{[γ_j^{\min},γ_j^{\max}]} (γ_j^{(k-1)} + ρ^{(k-1)} h_j(x^{(k)})) \text{for all} j=1,…,p
+γ_j^{(k)} =\operatorname{clip}_{[γ_{\min},γ_{\max}]} (γ_j^{(k-1)} + ρ^{(k-1)} h_j(x^{(k)})) \text{for all} j=1,…,p,
 ```
 and
 ```math
-λ_i^{(k)} =\operatorname{clip}_{[0,λ_i^{\max}]} (λ_i^{(k-1)} + ρ^{(k-1)} g_i(x^{(k)})) \text{for all}  i=1,…,m,
+λ_i^{(k)} =\operatorname{clip}_{[0,λ_{\max}]} (λ_i^{(k-1)} + ρ^{(k-1)} g_i(x^{(k)})) \text{for all}  i=1,…,m,
 ```
-where ``γ_j^{\min} \leq γ_j^{\max}`` and ``λ_i^{\max}`` are the multiplier boundaries. 
+where ``γ_{\min} \leq γ_{\max}`` and ``λ_{\max}`` are the multiplier boundaries. 
 
 Next, we update the accuracy tolerance ``ϵ`` by setting
 ```math
@@ -73,10 +73,9 @@ where ``θ_ρ \in (0,1)`` is a constant scaling factor.
 * `num_outer_itertgn` - (`30`)
 * `ϵ` - (`1e-3`) the accuracy tolerance
 * `ϵ_min` - (`1e-6`) the lower bound for the accuracy tolerance
-#### * `bound` - (`20`) a bound for the Lagrange multiplier
-* `γ_max` - (`20`)
-* `γ_min` - (`- γ_max`)
-* `λ_max` - (`20`)
+* `γ_max` - (`20.0`) an upper bound for the Lagrange multiplier belonging to the equality constraints
+* `γ_min` - (`- γ_max`) a lower bound for the Lagrange multiplier belonging to the equality constraints
+* `λ_max` - (`20.0`) an upper bound for the Lagrange multiplier belonging to the inequality constraints
 * `λ` - (`ones(len(`[`get_inequality_constraints`](@ref)`(p,x))`) the Lagrange multiplier with respect to the inequality constraints
 * `γ` - (`ones(len(`[`get_equality_constraints`](@ref)`(p,x))`) the Lagrange multiplier with respect to the equality constraints
 * `ρ` - (`1.0`) the penalty parameter
@@ -150,8 +149,9 @@ function augmented_Lagrangian_method!(
     num_outer_itertgn::Int=30,
     ϵ::Real=1e-3, #(starting)tolgradnorm
     ϵ_min::Real=1e-6, #endingtolgradnorm
-    bound::Real=20.0, 
-    #### multiplier boundaries anpassen
+    γ_max::Real=20.0,
+    γ_min::Real=-γ_max,
+    λ_max::Real=20.0,
     λ::Vector=ones(length(G(M,x))),
     γ::Vector=ones(length(H(M,x))),
     ρ::Real=1.0, 
@@ -168,19 +168,20 @@ function augmented_Lagrangian_method!(
         p,
         x,
         sub_problem,
-        sub_options,
-        max_inner_iter,
-        num_outer_itertgn,
-        ϵ,
-        ϵ_min,
-        bound,
-        #### multiplier boundaries anpassen
-        λ,
-        γ,
-        ρ,
-        τ,
[…]
