+++ src/solvers/augmented_Lagrangian_method.jl
@@ -60,15 +60,15 @@ where ``θ_ρ \in (0,1)`` is a constant scaling factor.
 * `M` – a manifold ``\mathcal M``
 * `F` – a cost function ``F:\mathcal M→ℝ`` to minimize
 * `gradF` – the gradient of the cost function
-* `sub_problem` – problem for the subsolver
-* `sub_options` – options of the subproblem
+
+# Optional
 * `G` – the inequality constraints
 * `H` – the equality constraints 
 * `gradG` – the gradient of the inequality constraints
 * `gradH` – the gradient of the equality constraints
-
-# Optional
 * `x` - initial point
+* `sub_problem` – problem for the subsolver
+* `sub_options` – options of the subproblem
 * `max_inner_iter` - (`200`) the maximum number of iterations the subsolver should perform in each iteration 
 * `num_outer_itertgn` - (`30`)
 * `ϵ` - (`1e-3`) the accuracy tolerance
@@ -94,19 +94,19 @@ OR
 function augmented_Lagrangian_method(
     M::AbstractManifold,
     F::TF,
-    gradF::TGF,
-    sub_problem::Problem,
-    sub_options::Options,
+    gradF::TGF;
     G::Function=x->[], 
     H::Function=x->[],
     gradG::Function=x->[],
-    gradH::Function=x->[];
+    gradH::Function=x->[],
     x=random_point(M), 
+    sub_problem::Problem = GradientProblem(M,F,gradF),
+    sub_options::Options = GradientDescentOptions(M,x),
     kwargs...,
 ) where {TF, TGF}
     x_res = allocate(x)
     copyto!(M, x_res, x)
-    return augmented_Lagrangian_method!(M, F, gradF, sub_problem, sub_options, G, H, gradG, gradH; x=x_res, kwargs...)
+    return augmented_Lagrangian_method!(M, F, gradF; G=G, H=H, gradG=gradG, gradH=gradH, x=x_res, sub_problem=sub_problem, sub_options=sub_options,kwargs...)
 end
 @doc raw"""
     augmented_Lagrangian_method!(M, F, gradF, sub_problem, sub_options, G, H, gradG, gradH)
@@ -125,8 +125,7 @@ where `M` is a Riemannian manifold, and ``f``, ``\{g_i\}_{i=1}^m`` and ``\{h_j\}
 * `M` – a manifold ``\mathcal M``
 * `F` – a cost function ``F:\mathcal M→ℝ`` to minimize
 * `gradF` – the gradient of the cost function
-* `sub_problem` – problem for the subsolver
-* `sub_options` – options of the subproblem
+## Optional 
 * `G` – the inequality constraints
 * `H` – the equality constraints 
 * `gradG` – the gradient of the inequality constraints
@@ -137,14 +136,14 @@ For more options, especially `x` for the initial point and `ρ` for the penalty 
 function augmented_Lagrangian_method!(
     M::AbstractManifold,
     F::TF,
-    gradF::TGF,
-    sub_problem::Problem,
-    sub_options::Options,
+    gradF::TGF;
     G::Function=x->[],
     H::Function=x->[],
     gradG::Function=x->[],
-    gradH::Function=x->[];
+    gradH::Function=x->[],
     x=random_point(M),
+    sub_problem::Problem = GradientProblem(M,F,gradF),
+    sub_options::Options = GradientDescentOptions(M,x),
     max_inner_iter::Int=200,
     num_outer_itertgn::Int=30,
     ϵ::Real=1e-3, #(starting)tolgradnorm
[…]
