### 2022-01-17 248a7fc
adds return_options, uses the maximum function for new_acc instead of max(), works on documentation
