### 2022-01-20 5cdf9f2
fixes the computation of the Lagrangian cost function
