+++ src/solvers/augmented_Lagrangian_method.jl
@@ -222,7 +222,7 @@ function step_solver!(p::ConstrainedProblem, o::ALMOptions, iter)
         cost_eq = 0
     end
     new_acc = max(maximum(abs.(max.(-o.λ./o.ρ, cost_ineq))), maximum(abs.(cost_eq)))
-    #new_acc = max(maximum(abs.(max.(-o.λ./o.ρ, Ref(cost_ineq)))), maximum(abs.(cost_eq))) ###kann man das Ref einfach so weglassen?
+    #new_acc = max(maximum(abs.(max.(-o.λ./o.ρ, Ref(cost_ineq)))), maximum(abs.(cost_eq))) ###what was Ref used for?
 
     # update ρ if necessary
     if iter == 1 || new_acc > o.τ * o.old_acc 
@@ -237,9 +237,27 @@ get_solver_result(o::ALMOptions) = o.x
 
 function get_Lagrangian_cost_function(p::ConstrainedProblem, o::ALMOptions)
     cost = x -> get_cost(p, x)
-    cost_ineq = x -> sum(max.(zeros(o.n_ineq), o.λ ./ o.ρ .+ get_inequality_constraints(p, x)))
-    cost_eq = x -> sum((get_equality_constraints(p, x) .+ o.γ./o.ρ)^2)
-    return (M,x) -> cost(x) + (o.ρ/2) * (cost_ineq(x) + cost_eq(x))
+    num_inequality_constraints = length(get_inequality_constraints(p,o.x))
+    num_equality_constraints = length(get_equality_constraints(p,o.x))
+    if num_inequality_constraints != 0 
+        cost_ineq = x -> sum(max.(zeros(num_inequality_constraints), o.λ ./ o.ρ .+ get_inequality_constraints(p, x)))
+    end
+    if num_equality_constraints != 0
+        cost_eq = x -> sum((get_equality_constraints(p, x) .+ o.γ./o.ρ).^2)
+    end
+    if num_inequality_constraints != 0
+        if num_equality_constraints != 0
+            return (M,x) -> cost(x) + (o.ρ/2) * (cost_ineq(x) + cost_eq(x))
+        else
+            return (M,x) -> cost(x) + (o.ρ/2) * cost_ineq(x)
+        end
+    else
+        if num_equality_constraints != 0
+            return (M,x) -> cost(x) + (o.ρ/2) * cost_eq(x)
+        else
+            return (M,x) -> cost(x) 
+        end
+    end
 end
 
 function get_Lagrangian_gradient_function(p::ConstrainedProblem, o::ALMOptions)
@@ -269,11 +287,3 @@ function get_Lagrangian_gradient_function(p::ConstrainedProblem, o::ALMOptions)
     end
 end
 
-# function get_Lagrangian_gradient_function(p::ConstrainedProblem, o::ALMOptions)
-#     grad = x -> get_gradient(p, x)
-#     grad_ineq = x -> sum(
-#         ((get_inequality_constraints(p, x) .* o.ρ .+ o.λ) .* get_grad_ineq(p, x)).*(get_inequality_constraints(p, x) .+ o.λ./o.ρ .>0)
-#         )
-#     grad_eq = x-> sum((get_equality_constraints(p, x) .* o.ρ .+ o.γ) .* get_grad_eq(p, x))
-#     return (M,x) -> grad(x) + grad_ineq(x) + grad_eq(x)
-# end
