### 2022-01-24 6d66c07
Update paper.md
Fix a typo in the orcid.
