### 2022-01-24 ff92602
Update paper.md
fix another copy paste error.
