### 2022-01-24 93b72f9
Update paper.md
Fixes another typo.
