### 2022-01-25 c7b3407
Fix random rotation tangent vectors (#100)
* Bump compat, fix random tangent for rotations, runs formatter.

* bump vesion number (leaving pi)
