### 2022-01-26 6a02730
Projection to the tangent space in the truncated conjugate gradient solver
The choice of the projection function is done through a named
argument "projection_to_tangent", which defaults to a function
doing nothing.
