### 2022-01-26 4c661f1
adds missing square in the computation of the Lagrangian cost function
