### 2022-01-28 57fbf32
correction of the lower boundary for gamma
