### 2022-02-01 63b4cb0
Fix Revise issue reported on Slack (#103)
* Fix Revise issue reported on Slack

Problem that occurs with this empty line:
```julia
julia> using Revise

julia> using Manifolds

julia> using Manopt
[ Info: Precompiling Manopt [0fc0a36d-df90-57f3-8f93-d78a9fc72bb5]
ERROR: InitError: BoundsError: attempt to access 3-element Vector{Any} at index [4]
Stacktrace:
  [1] getindex
    @ ./array.jl:861 [inlined]
  [2] iterate(iter::JuliaInterpreter.ExprSplitter, state::Nothing)
    @ JuliaInterpreter ~/.julia/packages/JuliaInterpreter/DxOBb/src/construct.jl:529
  [3] iterate
    @ ~/.julia/packages/Julia
[…]
