### 2022-02-07 185e9a0
Minor typos and grammar suggestions
