### 2022-02-07 #3866 @dpsanders
@kellertuer I sent a PR with some minor edits for typos and grammar suggestions.

Could you please add a couple of references for the applications listed in the Statement of Need, and to the ManifoldsBase.jl package.

Otherwise this seems ready!

Final steps:
- Please archive it, e.g. at Zenodo, and paste the URL here.  Please make sure the title and author info in Zenodo are *exactly* the same as in the paper.
- Record here the version of the software corresponding to the archived version.
