### 2022-02-07 #3866 @kellertuer
Oh, sorry for my impatience there, I was not expecting that you would also read it again :)

I just accepted and merged your patch.

Concerning the references – sure, I can add those later this week (hopefully tomorrow). Is it ok if I just link to ManifoldsBase.jl (and Manifolds.jl), for example do their package names as links to the repositories?

To the final steps
* I would merge the joss branch anyways and could just register a new version of Manopt.jl – I could also adapt the text to mention this version already. Would that fit in the sense of archiving the paper or should I archive the paper independently of the automated process (i.e. also _only_ the paper, so `paper.md` and `bibliography.bib`) ?
* sure then mentioning the software version here is not a problem.
