+++ joss/bibliography.bib
@@ -47,6 +47,15 @@
     TITLE  = {An Introduction to Optimization on Smooth Manifolds}
 }
 
+@inproceedings{GousenbourgerMassartMusolasAbsilJaquesHendrickxMarzouk:2017,
+    AUTHOR    = {Gousenbourger, Pierre-Yves and Massart, Estelle and Musolas, Antoni and Absil, P.-A. and Jacques, Laurent and Hendrickx, Julien M and Marzouk, Youssef},
+    BOOKTITLE = {ESANN2017},
+    PAGES     = {305--310},
+    PUBLISHER = {Springer},
+    TITLE     = {Piecewise-{B}{\'e}zier {$C^1$} smoothing on manifolds with application to wind field estimation},
+    YEAR      = {2017}
+}
+
 @article{HuangGallivanAbsil:2015:1,
     AUTHOR       = {Huang, W. and Gallivan, K. A. and Absil, P.-A.},
     DATE         = {2015},
@@ -58,6 +67,7 @@
     TITLE        = {A Broyden class of quasi-Newton methods for Riemannian optimization},
     VOLUME       = {25}
 }
+
 @online{AxenBaranBergmannRzecki:2021:1,
     AUTHOR     = {Axen, S. D. and Baran, M. and Bergmann, R. and Rzecki, K.},
     TITLE      = {Manifolds.jl: An Extensible {J}ulia Framework for Data Analysis on Manifolds},
@@ -92,6 +102,19 @@
     JOURNALTITLE = {Frontiers in Applied Mathematics and Statistics},
     TITLE        = {A variational model for data fitting on manifolds by minimizing the acceleration of a Bézier curve},
     VOLUME       = {4}
+}
+
+@article{BergmannChanHielscherPerschSteidl:2016,
+    NUMBER       = {2},
+    PAGES        = {281–304},
+    DOI          = {10.3934/ipi.2016001},
+    AUTHOR       = {Bergmann, R. and Chan, R. H. and Hielscher, R. and Persch, J. and Steidl, G.},
+    EPRINT       = {1505.07029},
+    YEAR         = {2016},
+    EPRINTTYPE   = {arXiv},
+    VOLUME       = {10},
+    JOURNALTITLE = {Inverse Problems and Imaging},
+    TITLE        = {Restoration of manifold-valued images by half-quadratic minimization},
 }
 
 @article{pymanopt,
@@ -182,4 +205,89 @@
     URL       = {https://doi.org/10.1007%2F978-3-030-80209-7_82},
     EDITOR    = {Nielsen F., Barbaresco F.},
     BOOKTITLE = {Geometric Science of Information. GSI 2021}
+}
+
+@article{VogtLellmann:2018:1,
+    AUTHOR       = {Vogt, Thomas and Lellmann, Jan},
+    DATE         = {2018},
+    DOI          = {10.1007/s10851-018-0827-8},
+    JOURNALTITLE = {Journal of Mathematical Imaging and Vision},
+    NUMBER       = {9},
+    PAGES        = {1482--1502},
+    TITLE        = {Measure-valued variational models with applications to diffusion-weighted imaging},
+    VOLUME       = {60}
+}
+
+@article{ValkonenBrediesKnoll2013,
+    DOI       = {10.1137/120867172},
+    YEAR      = 2013,
+    PUBLISHER = {Society for Industrial {\&} Applied Mathematics ({SIAM})},
+    VOLUME    = {6},
+    NUMBER    = {1},
+    PAGES     = {487--525},
+    AUTHOR    = {Tuomo Valkonen and Kristian Bredies and Florian Knoll},
+    TITLE     = {Total Generalized Variation in Diffusion Tensor Imaging},
[…]
