### 2022-02-08 13c52fd
Add reference to Manifolds.jl and ManifoldsBase.jl and fix two typos.
