+++ joss/bibliography.bib
@@ -43,7 +43,7 @@
 @booklet{Boumal:2020:1,
     AUTHOR = {Boumal, N.},
     URL    = {http://www.nicolasboumal.net/book},
-    DATE   = {2020-08},
+    DATE   = {2022},
     TITLE  = {An Introduction to Optimization on Smooth Manifolds}
 }
 
