### 2022-02-08 #3866 @kellertuer
I hope I finished most of your requests.

1. every application mentioned now also has a reference
2. I link to ManifoldsBase.jl on first occurrence (similar for Manifolds.jl) – but ManifoldsBase.jl does (not yet) have a zenodo DOI - we plan to do that with the next major release, which will also cover an own documentation (independent from Manifolds.jl) but this might still take a while.+

I already checked the result with whedon (the pre submit testy thing) and it looks finde, but I will trigger a build here as well.

So - is this ok? I am not sure what else to do with ManifoldsBase.jl

And – you wrote I should archive „it“ - just the paper? Or the whole software? The rules read like the software - but then the .zenodo file I have also covers everyone – even contributing minor functions – in said zenodo release.

However – I already prepared said version on the JOSS branch – merging that I will. tag 0.3.17 (which is now also the version mentioned in the paper.

So please give me a short feedback
1. ManifoldsBase.jl reference ok?
2. “just” 0.3.17 zenodo archive (automatic) ok or should I do a separate one just for the paper (the 0.3.17 will happen anyways)

For the whole software, we can surely do that by just merging the joss branch and doing a minor version bump (
