### 2022-02-09 #3866 @dpsanders
Thanks, the Zenodo archive I meant is for the software. So I think we're good, thanks.
