### 2022-02-09 bfce088
Runs formatter on the example.
