### 2022-02-09 84efaf0
Finalize JOSS Paper (#105)
* Initial version of JOSS paper.
* Minor typos and grammar suggestions (#104)

Co-authored-by: David P. Sanders <dpsanders@gmail.com>
