+++ joss/paper.md
@@ -1,0 +1,163 @@
+---
+title: 'Manopt.jl: Optimization on Manifolds in Julia'
+tags:
+  - Julia
+  - Riemannian manifolds
+  - optimization
+  - numerical analysis
+authors:
+  - name: Ronny Bergmann
+    orcid: 0000-0001-8342-7218
+    affiliation: 1
+affiliations:
+ - name: Norwegian University of Science and Technology, Department of Mathematical Sciences, Trondheim, Norway
+   index: 1
+date: 22 July 2021
+bibliography: bibliography.bib
+
+---
+
+# Summary
+
+[`Manopt.jl`](https://manoptjl.org) provides a set of optimization algorithms for optimization problems given on a Riemannian manifold $\mathcal M$.
+Based on a generic optimization framework, together with the interface [`ManifoldsBase.jl`](https://github.com/JuliaManifolds/ManifoldsBase.jl) for Riemannian manifolds, classical and recently developed methods are provided in an efficient implementation. Algorithms include the derivative-free Particle Swarm and Nelder–Mead algorithms, as well as classical gradient, conjugate gradient and stochastic gradient descent. Furthermore, quasi-Newton methods like a Riemannian L-BFGS [@HuangGallivanAbsil:2015:1] and nonsmooth optimization algorithms like a Cyclic Proximal Point Algorithm [@Bacak:2014:1], a (parallel) Douglas-Rachford algorithm [@BergmannPerschSteidl:2016:1] and a Chambolle-Pock algorithm [@BergmannHerzogSilvaLouzeiroTenbrinckVidalNunez:2021:1] are provided, together with several basic cost functions, gradients and proximal maps as well as debug and record capabilities.
+
+# Statement of Need
+
+In many applications and optimization tasks, non-linear data appears naturally.
+For example, when data on the sphere is measured [@GousenbourgerMassartMusolasAbsilJaquesHendrickxMarzouk:2017], diffusion data can be captured as a signal or even multivariate data of symmetric positive definite matrices [@ValkonenBrediesKnoll2013], and orientations like they appear for electron backscattered diffraction (EBSD) data [@BachmannHielscherSchaeben2011]. Another example are fixed rank matrices, appearing in matrix completion [@Vandereycken:2013:1].
+Working on these data, for example doing data interpolation and approximation [@BergmannGousenbourger:2018:2], denoising [@LellmannStrekalovskiyKoetterCremers:2013:1], [@BergmannFitschenPerschSteidl:2018], inpainting [@BergmannChanHielscherPerschSteidl:2016], or performing matrix completion [@GaoAbsil:2021], can usually be phrased as an optimization problem
+
+$$ \text{Minimize}\quad f(x) \quad \text{where } x\in\mathcal M, $$
+
+where the optimization problem is phrased on a Riemannian manifold $\mathcal M$.
+
+A main challenge of these algorithms is that, compared to the (classical) Euclidean case, there is no addition available. For example, on the unit sphere $\mathbb S^2$ of unit vectors in $\mathbb R^3$, adding two vectors of unit length yields a vector that is not of unit norm.
+The solution is to generalize the notion of a shortest path from the straight line to what is called a (shortest) geodesic, or acceleration-free curve.
+Similarly, other features and properties also have to be rephrased and generalized when performing optimization on a Riemannian manifold.
+Algorithms to perform the optimization can still often be stated in a generic way, i.e. on an arbitrary Riemannian manifold $\mathcal M$.
+Further examples and a thorough introduction can be found in [@AbsilMahonySepulchre:2008:1; @Boumal:2020:1].
+
[…]
