### 2022-02-09 #3866 @kellertuer
Supert! I just merged the JOSS branch (it still exists of course to not confuse whedon) and registered that as a new version.

the new DOI including the paper is [10.5281/zenodo.6019027](https://doi.org/10.5281/zenodo.6019027), it includes the paper as well as a script file for the example done in the paper – and this example also recreates the figure if run.

Thanks for all your help :)
