### 2022-02-10 #3866 @kthyng
Hi @kellertuer! a few items:

1. version and zenodo look good ✅
2. paper:
- statement of need, first paragraph: you have a list of references but each are in their own set of parentheses. Put the brackets on the outside of the list, with ";" between, to get one set of parentheses on the outside. [docs](@whedon generate pdf from branch joss-paper)
- functionality, paragraph 2: "see e.g. (..." but the reference should not be parenthetical. Just remove the [] around the reference.
- other references cited inline I see should not be parenthetical sometimes — remove the brackets when they should be inline.
- Preserve capitalization in your references by putting {} around characters in your .bib file. For example, Absil et al has
"riemannian" not capitalized. Check the other references too.
