### 2022-02-10 #3866 @kellertuer
Thanks for the tipps

* I actually only learned about `;` in cites yesterday and had hoped I changed all places – seem to have missed the one in the fist paragraph - thanks, fixed that
* oh, I was not aware of the inline cite. I adapted the mention cite and all other places where the text was using online cites – they even work with ; - nice!
* sorry for the capitalisation errors, I was a little uncareful yesterday I think.
* While going through capitalisation I even notices I had one reference double-included and reduced that to only once having that in the bib-file.
