+++ joss/bibliography.bib
@@ -13,7 +13,7 @@
     JOURNALTITLE = {SIAM Journal on Optimization},
     NUMBER       = {3},
     PAGES        = {1542--1566},
-    TITLE        = {Computing medians and means in Hadamard spaces},
+    TITLE        = {Computing medians and means in {H}adamard spaces},
     VOLUME       = {24}
 }
 
@@ -23,7 +23,7 @@
     DATE         = {2021-01},
     DOI          = {10.1007/s10208-020-09486-5},
     JOURNALTITLE = {Foundations of Computational Mathematics},
-    TITLE        = {Fenchel duality theory and a primal-dual algorithm on Riemannian manifolds}
+    TITLE        = {{F}enchel duality theory and a primal-dual algorithm on {R}iemannian manifolds}
 }
 
 @article{BergmannPerschSteidl:2016:1,
@@ -35,7 +35,7 @@
     JOURNALTITLE = {SIAM Journal on Imaging Sciences},
     NUMBER       = {4},
     PAGES        = {901--937},
-    TITLE        = {A parallel Douglas Rachford algorithm for minimizing ROF-like functionals on images with values in symmetric Hadamard manifolds},
+    TITLE        = {A parallel {D}ouglas–{R}achford algorithm for minimizing {ROF}-like functionals on images with values in symmetric {H}adamard manifolds},
     VOLUME       = {9}
 }
 
@@ -56,21 +56,9 @@
     YEAR      = {2017}
 }
 
-@article{HuangGallivanAbsil:2015:1,
-    AUTHOR       = {Huang, W. and Gallivan, K. A. and Absil, P.-A.},
-    DATE         = {2015},
-    DOI          = {10.1137/140955483},
-    ISSN         = {1052-6234},
-    JOURNALTITLE = {SIAM Journal on Optimization},
-    NUMBER       = {3},
-    PAGES        = {1660--1685},
-    TITLE        = {A Broyden class of quasi-Newton methods for Riemannian optimization},
-    VOLUME       = {25}
-}
-
 @online{AxenBaranBergmannRzecki:2021:1,
     AUTHOR     = {Axen, S. D. and Baran, M. and Bergmann, R. and Rzecki, K.},
-    TITLE      = {Manifolds.jl: An Extensible {J}ulia Framework for Data Analysis on Manifolds},
+    TITLE      = {{M}anifolds.jl: An Extensible {J}ulia Framework for Data Analysis on Manifolds},
     YEAR       = {2021},
     EPRINT     = {2106.08777},
     EPRINTTYPE = {arXiv}
@@ -119,7 +107,7 @@
 
 @article{pymanopt,
     AUTHOR  = {Townsend, T. and Koep, N. and Weichwald, S.},
-    TITLE   = {Pymanopt: A Python Toolbox for Optimization on Manifolds using Automatic Differentiation},
+    TITLE   = {{P}ymanopt: A Python Toolbox for Optimization on Manifolds using Automatic Differentiation},
     JOURNAL = {Journal of Machine Learning Research},
     YEAR    = {2016},
     VOLUME  = {17},
@@ -156,7 +144,7 @@
     JOURNALTITLE = {Association for Computing Machinery. Transactions on Mathematical Software},
     NUMBER       = {4},
     PAGES        = {Art. 43,21},
-    TITLE        = {ROPTLIB: an object-oriented {C++} library for optimization on Riemannian manifolds},
+    TITLE        = {ROPTLIB: an object-oriented {C++} library for optimization on {R}iemannian manifolds},
     VOLUME       = {44}
 }
 
@@ -167,7 +155,7 @@
     JOURNALTITLE = {SIAM Journal on Optimization},
[…]
